Cardano.Ledger.Core

Description

This module defines core type families which we know to vary from era to era.

Families in this module should be indexed on era.

It is intended for qualified import: > import qualified Cardano.Ledger.Core as Core

Era-changing types

type family TxBody era ∷ Type

The body of a transaction.

type family Value era ∷ Type

A value is something which quantifies a transaction output.

type family Script era ∷ Type

Scripts which may lock transaction outputs in this era

Constraint synonyms

type ChainData t = (Eq t, Show t, NoThunks t, Typeable t)

Common constraints

NOTE: Ord is not included, as Ord for a Block or a NewEpochState doesn't make sense.

type SerialisableData t = (FromCBOR t, ToCBOR t)

Constraints for serialising from/to CBOR

type AnnotatedData t = (FromCBOR (Annotator t), ToCBOR t)

Constraints for serialising from/to CBOR using Annotator