Cardano.Ledger.Shelley.API.Validation

Description

Interface to the block validation and chain extension logic in the Shelley API.

Documentation

type ShelleyState = NewEpochState

Type alias for the state updated by TICK and BBODY rules

data TickTransitionError crypto

data BlockTransitionError crypto

applyTickTransition ∷ ∀ crypto. Crypto crypto ⇒ Globals → ShelleyState crypto → SlotNo → ShelleyState crypto

Apply the header level ledger transition.

This handles checks and updates that happen on a slot tick, as well as a few header level checks, such as size constraints.

applyBlockTransition ∷ ∀ crypto m. (Crypto crypto, MonadError (BlockTransitionError crypto) m, Signable (DSIGN crypto) (TxBody crypto)) ⇒ Globals → ShelleyState crypto → Block crypto → m (ShelleyState crypto)

Apply the block level ledger transition.