Cardano.Ledger.Shelley.API.Wallet

Documentation

getNonMyopicMemberRewards ∷ Globals → ShelleyState crypto → Set (Credential crypto) → Map (Credential crypto) (Map (KeyHash crypto) Coin)

Calculate the Non-Myopic Pool Member Rewards for a set of credentials. For each given credential, this function returns a map from each stake pool (identified by the key hash of the pool operator) to the non-myopic pool member reward for that stake pool.