Cardano.Ledger.Val

Description

This module defines a generalised notion of a "value" - that is, something with which we may quantify a transaction output.

This module is intended for qualified import: > import qualified Cardano.Ledger.Val as Val

Documentation

class (Abelian t, Eq t, PartialOrd t, Show t, Typeable t, NFData t, NoUnexpectedThunks t, ToCBOR t, FromCBOR t) ⇒ Val t where

Minimal complete definition

scale, coin, inject, size

Methods

scale ∷ Integral i ⇒ i → t → t

Multiply the value by a scalar

isZero ∷ t → Bool

Is the argument zero?

coin ∷ t → Coin

Get the ADA present in the value (since ADA is our "blessed" currency)

inject ∷ Coin → t

Create a value containing only this amount of ADA

size ∷ t → Integer

(~~) ∷ Group g ⇒ g → g → g

Group subtraction. When we move to groups-0.5 we can export this from there.

scaledMinDeposit ∷ Val v ⇒ v → Coin → Coin

Re-exports

invert ∷ Group m ⇒ m → m

(<=) ∷ PartialOrd a ⇒ a → a → Bool

(>=) ∷ PartialOrd a ⇒ a → a → Bool

(==) ∷ PartialOrd a ⇒ a → a → Bool

(/=) ∷ PartialOrd a ⇒ a → a → Bool

(>) ∷ PartialOrd a ⇒ a → a → Bool

(<) ∷ PartialOrd a ⇒ a → a → Bool

compare ∷ PartialOrd a ⇒ a → a → Maybe Ordering