Shelley.Spec.Ledger.API.ByronTranslation

Documentation

mkInitialShelleyLedgerView ∷ ∀ c. ShelleyGenesis (ShelleyEra c) → LedgerView c

We construct a LedgerView using the Shelley genesis config in the same way as translateToShelleyLedgerState.

translateToShelleyLedgerState ∷ ∀ c. (Crypto c, ADDRHASH c ~ Blake2b_224) ⇒ ShelleyGenesis (ShelleyEra c) → EpochNo → ChainValidationState → NewEpochState (ShelleyEra c)

Exported for testing purposes

translateCompactTxOutByronToShelley ∷ CompactTxOut → TxOut (ShelleyEra c)

translateTxIdByronToShelley ∷ (Crypto c, ADDRHASH c ~ Blake2b_224) ⇒ TxId → TxId (ShelleyEra c)

We use the same hashing algorithm so we can unwrap and rewrap the bytes. We don't care about the type that is hashed, which will differ going from Byron to Shelley, we just use the hashes as IDs.