Shelley.Spec.Ledger.API.Mempool

Description

Interface to the Shelley ledger for the purposes of managing a Shelley mempool.

Documentation

type MempoolEnv = LedgersEnv

type MempoolState = LedgerState

mkMempoolEnv ∷ ShelleyState crypto → SlotNo → MempoolEnv

Construct the environment used to validate transactions from the full ledger state.

Note that this function also takes a slot. During slot validation, the slot given here is the slot of the block containing the transactions. This slot is used for quite a number of things, but in general these do not determine the validity of the transaction. There are two exceptions:

Each transaction has a ttl (time-to-live) value. If the slot is beyond this value, then the transaction is invalid.

If the transaction contains a protocol update proposal, then it may only be included until a certain number of slots before the end of the epoch. A protocol update proposal submitted after this is considered invalid.

mkMempoolState ∷ ShelleyState crypto → MempoolState crypto

Construct a mempool state from the wider ledger state.

The given mempool state may then be evolved using applyTxs, but should be regenerated when the ledger state gets updated (e.g. through application of a new block).

data ApplyTxError crypto

Constructors

ApplyTxError [PredicateFailure (LEDGERS crypto)]

applyTxs ∷ ∀ crypto m. (Crypto crypto, MonadError (ApplyTxError crypto) m, Signable (DSIGN crypto) (TxBody crypto)) ⇒ Globals → MempoolEnv → Seq (Tx crypto) → MempoolState crypto → m (MempoolState crypto)

overShelleyState ∷ Applicative f ⇒ (MempoolState c → f (MempoolState c)) → ShelleyState c → f (ShelleyState c)

Transform a function over mempool states to one over the full Shelley state.