Shelley.Spec.Ledger.API.Protocol

Description

Integration between the Shelley ledger and its corresponding (Transitional Praos) protocol.

In particular, this code supports extracting the components of the ledger state needed for protocol execution, both now and in a 2k-slot window.

Documentation

class (Crypto c, DSignable c (OCertSignable c), DSignable c (Hash c EraIndependentTxBody), KESignable c (BHBody c), VRFSignable c Seed) ⇒ PraosCrypto c

class (ChainData (ChainDepState (Crypto era)), SerialisableData (ChainDepState (Crypto era)), Eq (ChainTransitionError (Crypto era)), Show (ChainTransitionError (Crypto era)), Show (LedgerView (Crypto era)), Show (FutureLedgerViewError era)) ⇒ GetLedgerView era where

Minimal complete definition

Nothing

Methods

currentLedgerView ∷ NewEpochState era → LedgerView (Crypto era)

futureLedgerView ∷ MonadError (FutureLedgerViewError era) m ⇒ Globals → NewEpochState era → SlotNo → m (LedgerView (Crypto era))

default futureLedgerView ∷ (ShelleyBased era, MonadError (FutureLedgerViewError era) m) ⇒ Globals → NewEpochState era → SlotNo → m (LedgerView (Crypto era))

data LedgerView crypto

Data required by the Transitional Praos protocol from the Shelley ledger.

Constructors

LedgerView

Fields

lvD ∷ UnitInterval

lvExtraEntropy ∷ Nonce

lvPoolDistr ∷ PoolDistr crypto

lvGenDelegs ∷ GenDelegs crypto

lvChainChecks ∷ ChainChecksData

newtype FutureLedgerViewError era

Constructors

FutureLedgerViewError [PredicateFailure (TICKF era)]

Chain state operations

The chain state is an amalgam of the protocol state and the ticked nonce.

data ChainDepState crypto

Constructors

ChainDepState

Fields

csProtocol ∷ !(PrtclState crypto)

csTickn ∷ !TicknState

csLabNonce ∷ !Nonce

Nonce constructed from the hash of the last applied block header.

newtype ChainTransitionError crypto

Constructors

ChainTransitionError [PredicateFailure (PRTCL crypto)]

tickChainDepState

Arguments

∷ Globals

→ LedgerView crypto

→ Bool
Are we in a new epoch?

→ ChainDepState crypto

→ ChainDepState crypto

Tick the chain state to a new epoch.

updateChainDepState ∷ ∀ crypto m. (PraosCrypto crypto, MonadError (ChainTransitionError crypto) m) ⇒ Globals → LedgerView crypto → BHeader crypto → ChainDepState crypto → m (ChainDepState crypto)

Update the chain state based upon a new block header.

This also updates the last applied block hash.

reupdateChainDepState ∷ ∀ crypto. PraosCrypto crypto ⇒ Globals → LedgerView crypto → BHeader crypto → ChainDepState crypto → ChainDepState crypto

Re-update the chain state based upon a new block header.

This function does no validation of whether the header is internally valid or consistent with the chain it is being applied to; the caller must ensure that this is valid through having previously applied it.