Shelley.Spec.Ledger.API.Types

Documentation

hashVerKeyVRF ∷ (VRFAlgorithm v, HashAlgorithm h) ⇒ VerKeyVRF v → Hash h (VerKeyVRF v)

newtype Coin

The amount of value held by a transaction output.

Constructors

Coin

Fields

unCoin ∷ Integer

word64ToCoin ∷ Word64 → Coin

type VerKeyVRF c = VerKeyVRF (VRF c)

type SignKeyVRF c = SignKeyVRF (VRF c)

type CertifiedVRF c = CertifiedVRF (VRF c)

type VerKeyKES c = VerKeyKES (KES c)

type SignKeyKES c = SignKeyKES (KES c)

type SignedKES c = SignedKES (KES c)

type SignKeyDSIGN c = SignKeyDSIGN (DSIGN c)

type SignedDSIGN c = SignedDSIGN (DSIGN c)

type Hash c = Hash (HASH c)

newtype GenDelegs crypto

Constructors

GenDelegs

Fields

unGenDelegs ∷ Map (KeyHash 'Genesis crypto) (GenDelegPair crypto)

data GenDelegPair crypto

Constructors

GenDelegPair

Fields

genDelegKeyHash ∷ !(KeyHash 'GenesisDelegate crypto)

genDelegVrfHash ∷ !(Hash crypto (VerKeyVRF crypto))

type KESignable c = Signable (KES c)

newtype KeyHash (discriminator ∷ KeyRole) crypto

Discriminated hash of public Key

Constructors

KeyHash (Hash (ADDRHASH crypto) (VerKeyDSIGN (DSIGN crypto)))

data KeyPair (kd ∷ KeyRole) crypto

Pair of signing key and verification key, with a usage role.

Constructors

KeyPair

Fields

vKey ∷ !(VKey kd crypto)

sKey ∷ !(SignKeyDSIGN (DSIGN crypto))

newtype VKey (kd ∷ KeyRole) crypto

Discriminated verification key

We wrap the basic VerKeyDSIGN in order to add the key role.

Constructors

VKey

Fields

unVKey ∷ VerKeyDSIGN (DSIGN crypto)

coerceKeyRole ∷ HasKeyRole a ⇒ a r crypto → a r' crypto

General coercion of key roles.

The presence of this function is mostly to help the user realise where they are converting key roles.

data KeyRole

The role of a key.

Note that a role is not _fixed_, nor is it unique. In particular, keys may variously be used as witnesses, and so in many case we will change the role of a key to the Witness role.

It is also perfectly allowable for a key to be used in many roles; there is nothing prohibiting somebody using the same underlying key as their payment and staking key, as well as the key for their stake pool. So these roles are more intended for two purposes:

To make explicit how we are using a key in the specifications

To provide a guide to downstream implementors, for whom the profusion of keys may be confusing.

Constructors

Genesis

GenesisDelegate

Payment

Staking

StakePool

BlockIssuer

Witness

hashKey ∷ Crypto crypto ⇒ VKey kd crypto → KeyHash kd crypto

Hash a given public key

data MetaData where

Bundled Patterns

pattern MetaData ∷ Map Word64 MetaDatum → MetaData

data MetaDatum

A generic metadatum type.

Constructors

Map [(MetaDatum, MetaDatum)]

List [MetaDatum]

I !Integer

B !ByteString

S !Text

data Network

Constructors

Testnet

Mainnet

data Globals

Constructors

Globals

Fields

epochInfo ∷ !(EpochInfo Identity)

slotsPerKESPeriod ∷ !Word64

stabilityWindow ∷ !Word64

The window size in which our chosen chain growth property guarantees at least k blocks. From the paper "Ouroboros praos: An adaptively-secure, semi-synchronous proof-of-stake protocol". The stabilityWindow constant is used in a number of places; for example, protocol updates must be submitted at least twice this many slots before an epoch boundary.

randomnessStabilisationWindow ∷ !Word64

Number of slots before the end of the epoch at which we stop updating the candidate nonce for the next epoch.

securityParameter ∷ !Word64

Maximum number of blocks we are allowed to roll back

maxKESEvo ∷ !Word64

Maximum number of KES iterations

quorum ∷ !Word64

Quorum for update system votes and MIR certificates

maxMajorPV ∷ !Natural

All blocks invalid after this protocol version

maxLovelaceSupply ∷ !Word64

Maximum number of lovelace in the system

activeSlotCoeff ∷ !ActiveSlotCoeff

Active Slot Coefficient, named f in "Ouroboros Praos: An adaptively-secure, semi-synchronous proof-of-stake protocol"

networkId ∷ !Network

The network ID

newtype Port

Constructors

Port

Fields

portToWord16 ∷ Word16

data StrictMaybe a

Strict Maybe.

TODO move to cardano-prelude

Constructors

SNothing

SJust !a

data Nonce

Evolving nonce type.

Constructors

Nonce !(Hash Blake2b_256 Nonce)

NeutralNonce
Identity element

newtype ScriptHash era

Constructors

ScriptHash (Hash (ADDRHASH (Crypto era)) (Script era))

data MultiSig era where

Bundled Patterns

pattern RequireMOf ∷ Era era ⇒ Int → [MultiSig era] → MultiSig era

pattern RequireSignature ∷ Era era ⇒ KeyHash 'Witness (Crypto era) → MultiSig era

pattern RequireAllOf ∷ Era era ⇒ [MultiSig era] → MultiSig era

pattern RequireAnyOf ∷ Era era ⇒ [MultiSig era] → MultiSig era

data TicknPredicateFailure

data TicknState

Constructors

TicknState

Fields

ticknStateEpochNonce ∷ !Nonce

ticknStatePrevHashNonce ∷ !Nonce

data TicknEnv

Constructors

TicknEnv

Fields

ticknEnvExtraEntropy ∷ Nonce

ticknEnvCandidateNonce ∷ Nonce

ticknEnvHashHeaderNonce ∷ Nonce

Hash of the last header of the previous epoch as a nonce.

data TICKN

data BootstrapWitness era where

Bundled Patterns

pattern BootstrapWitness ∷ Era era ⇒ VKey 'Witness (Crypto era) → SignedDSIGN (Crypto era) (Hash (Crypto era) EraIndependentTxBody) → ChainCode → ByteString → BootstrapWitness era

data OBftSlot crypto

Constructors

NonActiveSlot

ActiveSlot !(KeyHash 'Genesis crypto)

isOverlaySlot ∷ SlotNo → UnitInterval → SlotNo → Bool

classifyOverlaySlot ∷ SlotNo → Set (KeyHash 'Genesis crypto) → UnitInterval → ActiveSlotCoeff → SlotNo → OBftSlot crypto

lookupInOverlaySchedule ∷ SlotNo → Set (KeyHash 'Genesis crypto) → UnitInterval → ActiveSlotCoeff → SlotNo → Maybe (OBftSlot crypto)

newtype ProposedPPUpdates era

Update operation for protocol parameters structure @PParams

Constructors

ProposedPPUpdates (Map (KeyHash 'Genesis (Crypto era)) (PParamsUpdate era))

data Update era

Update Proposal

Constructors

Update !(ProposedPPUpdates era) !EpochNo

data ProtVer

Constructors

ProtVer

Fields

pvMajor ∷ !Natural

pvMinor ∷ !Natural

type PParams era = PParams' Identity era

data PParams' f era

Protocol parameters.

We use the HKD type family so that the protocol parameters type and the type for the updates to the protocol parameters can share records fields. The protocol parameters will have type PParams' Identity, and the updates will have type PParams' StrictMaybe, though Identity will be hidden from use.

For example:

  myParameters =
    PParams
      { _minfeeA = 0,
        _minfeeB = 0,
        ...
      }

  myUpdate =
    PParamsUpdate
      { _minfeeA = SNothing,
        _minfeeB = SJust 42,
        ...
      }

Constructors

PParams

Fields

_minfeeA ∷ !(HKD f Natural)

The linear factor for the minimum fee calculation

_minfeeB ∷ !(HKD f Natural)

The constant factor for the minimum fee calculation

_maxBBSize ∷ !(HKD f Natural)

Maximal block body size

_maxTxSize ∷ !(HKD f Natural)

Maximal transaction size

_maxBHSize ∷ !(HKD f Natural)

Maximal block header size

_keyDeposit ∷ !(HKD f Coin)

The amount of a key registration deposit

_poolDeposit ∷ !(HKD f Coin)

The amount of a pool registration deposit

_eMax ∷ !(HKD f EpochNo)

epoch bound on pool retirement

_nOpt ∷ !(HKD f Natural)

Desired number of pools

_a0 ∷ !(HKD f Rational)

Pool influence

_rho ∷ !(HKD f UnitInterval)

Monetary expansion

_tau ∷ !(HKD f UnitInterval)

Treasury expansion

_d ∷ !(HKD f UnitInterval)

Decentralization parameter

_extraEntropy ∷ !(HKD f Nonce)

Extra entropy

_protocolVersion ∷ !(HKD f ProtVer)

Protocol version

_minUTxOValue ∷ !(HKD f Coin)

Minimum UTxO value

_minPoolCost ∷ !(HKD f Coin)

Minimum Stake Pool Cost

data OCert crypto

Constructors

OCert

Fields

ocertVkHot ∷ !(VerKeyKES crypto)

The operational hot key

ocertN ∷ !Word64

counter

ocertKESPeriod ∷ !KESPeriod

Start of key evolving signature period

ocertSigma ∷ !(SignedDSIGN crypto (OCertSignable crypto))

Signature of block operational certificate content

newtype KESPeriod

Constructors

KESPeriod

Fields

unKESPeriod ∷ Word

data OCertEnv crypto

Constructors

OCertEnv

Fields

ocertEnvStPools ∷ Set (KeyHash 'StakePool crypto)

ocertEnvGenDelegs ∷ Set (KeyHash 'GenesisDelegate crypto)

data Ptr

Pointer to a slot, transaction index and index in certificate list.

Constructors

Ptr !SlotNo !Ix !Ix

data StakeReference era

Constructors

StakeRefBase !(StakeCredential era)

StakeRefPtr !Ptr

StakeRefNull

data Credential (kr ∷ KeyRole) era

Script hash or key hash for a payment or a staking object.

Note that credentials (unlike raw key hashes) do appear to vary from era to era, since they reference the hash of a script, which can change. This parameter is a phantom, however, so in actuality the instances will remain the same.

Constructors

ScriptHashObj !(ScriptHash era)

KeyHashObj !(KeyHash kr (Crypto era))

data RewardAcnt era

An account based address for rewards

Constructors

RewardAcnt

Fields

getRwdNetwork ∷ !Network

getRwdCred ∷ !(Credential 'Staking era)

data Addr era

An address for UTxO.

Constructors

Addr Network (PaymentCredential era) (StakeReference era)

AddrBootstrap (BootstrapAddress era)

computeStabilityWindow ∷ Word64 → ActiveSlotCoeff → Word64

Calculate the stability window (e.g. the number of slots needed for a block to become stable) from the security param and the active slot coefficient.

The value 3k/f is determined to be a suitabe value as per https://docs.google.com/document/d/1B8BNMx8jVWRjYiUBOaI3jfZ7dQNvNTSDODvT5iOuYCU/edit#heading=h.qh2zcajmu6hm

computeRandomnessStabilisationWindow ∷ Word64 → ActiveSlotCoeff → Word64

Calculate the randomness stabilisation window from the security param and the active slot coefficient.

The value 4k/f is determined to be a suitabe value as per https://docs.google.com/document/d/1B8BNMx8jVWRjYiUBOaI3jfZ7dQNvNTSDODvT5iOuYCU/edit#heading=h.qh2zcajmu6hm

newtype StakeCreds era

Constructors

StakeCreds

Fields

unStakeCreds ∷ Map (Credential 'Staking era) SlotNo

data WitVKey kr era where

Proof/Witness that a transaction is authorized by the given key holder.

Bundled Patterns

pattern WitVKey ∷ (Typeable kr, Era era) ⇒ VKey kr (Crypto era) → SignedDSIGN (Crypto era) (Hash (Crypto era) EraIndependentTxBody) → WitVKey kr era

newtype TxBody era

Constructors

TxBodyY (MemoBytes (TxBodyX era))

Bundled Patterns

pattern TxBody ∷ ProperTo era ⇒ Set (TxIn era) → StrictSeq (TxOut era) → StrictSeq (DCert era) → Wdrl era → Coin → SlotNo → StrictMaybe (Update era) → StrictMaybe (MetaDataHash era) → TxBody era
Pattern for use by external users

data DCert era

A heavyweight certificate.

Constructors

DCertDeleg !(DelegCert era)

DCertPool !(PoolCert era)

DCertGenesis !(GenesisDelegCert era)

DCertMir !(MIRCert era)

data MIRCert era

Move instantaneous rewards certificate

Constructors

MIRCert

Fields

mirPot ∷ MIRPot

mirRewards ∷ Map (Credential 'Staking era) Coin

data MIRPot

Constructors

ReservesMIR

TreasuryMIR

data GenesisDelegCert era

Genesis key delegation certificate

Constructors

GenesisDelegCert !(KeyHash 'Genesis (Crypto era)) !(KeyHash 'GenesisDelegate (Crypto era)) !(Hash (Crypto era) (VerKeyVRF (Crypto era)))

data PoolCert era

Constructors

RegPool !(PoolParams era)
A stake pool registration certificate.

RetirePool !(KeyHash 'StakePool (Crypto era)) !EpochNo
A stake pool retirement certificate.

data DelegCert era

Constructors

RegKey !(StakeCredential era)
A stake key registration certificate.

DeRegKey !(StakeCredential era)
A stake key deregistration certificate.

Delegate !(Delegation era)
A stake delegation certificate.

data TxOut era

The output of a UTxO.

Constructors

TxOutCompact !(CompactAddr era) !(CompactForm (Value era))

Bundled Patterns

pattern TxOut ∷ ShelleyBased era ⇒ Addr era → Value era → TxOut era

data TxIn era

The input of a UTxO.

Constructors

TxInCompact !(TxId era) !Word64

Bundled Patterns

pattern TxIn ∷ Era era ⇒ TxId era → Natural → TxIn era

newtype TxId era

A unique ID of a transaction, which is computable from the transaction.

Constructors

TxId

Fields

_unTxId ∷ Hash (Crypto era) EraIndependentTxBody

newtype Wdrl era

Constructors

Wdrl

Fields

unWdrl ∷ Map (RewardAcnt era) Coin

data PoolParams era

A stake pool.

Constructors

PoolParams

Fields

_poolId ∷ !(KeyHash 'StakePool (Crypto era))

_poolVrf ∷ !(Hash (Crypto era) (VerKeyVRF (Crypto era)))

_poolPledge ∷ !Coin

_poolCost ∷ !Coin

_poolMargin ∷ !UnitInterval

_poolRAcnt ∷ !(RewardAcnt era)

_poolOwners ∷ !(Set (KeyHash 'Staking (Crypto era)))

_poolRelays ∷ !(StrictSeq StakePoolRelay)

_poolMD ∷ !(StrictMaybe PoolMetaData)

data StakePoolRelay

Constructors

SingleHostAddr !(StrictMaybe Port) !(StrictMaybe IPv4) !(StrictMaybe IPv6)
One or both of IPv4 & IPv6

SingleHostName !(StrictMaybe Port) !DnsName
An A or AAAA DNS record

MultiHostName !DnsName
A SRV DNS record

data PoolMetaData

Constructors

PoolMetaData

Fields

_poolMDUrl ∷ !Url

_poolMDHash ∷ !ByteString

data Delegation era

The delegation of one stake key to another.

Constructors

Delegation

Fields

_delegator ∷ !(StakeCredential era)

_delegatee ∷ !(KeyHash 'StakePool (Crypto era))

data Tx era

A fully formed transaction.

Constructors

Tx' !(TxBody era) !(WitnessSet era) !(StrictMaybe MetaData) ByteString

Bundled Patterns

pattern Tx ∷ TxBodyConstraints era ⇒ TxBody era → WitnessSet era → StrictMaybe MetaData → Tx era

type WitnessSet = WitnessSetHKD Identity

individualPoolStake ∷ IndividualPoolStake crypto → Rational

newtype PoolDistr crypto

Constructors

PoolDistr

Fields

unPoolDistr ∷ Map (KeyHash 'StakePool crypto) (IndividualPoolStake crypto)

newtype UTxO era

The unspent transaction outputs.

Constructors

UTxO

Fields

unUTxO ∷ Map (TxIn era) (TxOut era)

balance ∷ ShelleyBased era ⇒ UTxO era → Value era

Determine the total balance contained in the UTxO.

data ValidationErr

Constructors

EpochNotLongEnough EpochSize Word64 Rational EpochSize

MaxKESEvolutionsUnsupported Word64 Word

QuorumTooSmall Word64 Word64 Word64

data ShelleyGenesis era

Shelley genesis information

Note that this is needed only for a pure Shelley network, hence it being defined here rather than in its own module. In mainnet, Shelley will transition naturally from Byron, and thus will never have its own genesis information.

Constructors

ShelleyGenesis

Fields

sgSystemStart ∷ !UTCTime

sgNetworkMagic ∷ !Word32

sgNetworkId ∷ !Network

sgActiveSlotsCoeff ∷ !Rational

sgSecurityParam ∷ !Word64

sgEpochLength ∷ !EpochSize

sgSlotsPerKESPeriod ∷ !Word64

sgMaxKESEvolutions ∷ !Word64

sgSlotLength ∷ !NominalDiffTime

sgUpdateQuorum ∷ !Word64

sgMaxLovelaceSupply ∷ !Word64

sgProtocolParams ∷ !(PParams era)

sgGenDelegs ∷ !(Map (KeyHash 'Genesis (Crypto era)) (GenDelegPair (Crypto era)))

sgInitialFunds ∷ !(Map (Addr era) Coin)

sgStaking ∷ !(ShelleyGenesisStaking era)

data ShelleyGenesisStaking era

Genesis Shelley staking configuration.

This allows us to configure some initial stake pools and delegation to them, in order to test Praos in a static configuration, without requiring on-chain registration and delegation.

For simplicity, pools defined in the genesis staking do not pay deposits for their registration.

Constructors

ShelleyGenesisStaking

Fields

sgsPools ∷ !(Map (KeyHash 'StakePool (Crypto era)) (PoolParams era))

Pools to register

The key in this map is the hash of the public key of the _pool_. This need not correspond to any payment or staking key, but must correspond to the cold key held by TPraosIsCoreNode.

sgsStake ∷ !(Map (KeyHash 'Staking (Crypto era)) (KeyHash 'StakePool (Crypto era)))

Stake-holding key hash credentials and the pools to delegate that stake to. We require the raw staking key hash in order to:

Avoid pointer addresses, which would be tricky when there's no slot or transaction to point to.

Avoid script credentials.

emptyGenesisStaking ∷ ShelleyGenesisStaking era

Empty genesis staking

sgActiveSlotCoeff ∷ ShelleyGenesis era → ActiveSlotCoeff

genesisUtxO ∷ ShelleyBased era ⇒ ShelleyGenesis era → UTxO era

initialFundsPseudoTxIn ∷ ∀ era. Era era ⇒ Addr era → TxIn era

Compute the TxIn of the initial UTxO pseudo-transaction corresponding to the given address in the genesis initial funds.

The Shelley initial UTxO is constructed from the sgInitialFunds which is not a full UTxO but just a map from addresses to coin values.

This gets turned into a UTxO by making a pseudo-transaction for each address, with the 0th output being the coin value. So to spend from the initial UTxO we need this same TxIn to use as an input to the spending transaction.

describeValidationErr ∷ ValidationErr → Text

validateGenesis ∷ ∀ era. Era era ⇒ ShelleyGenesis era → Either [ValidationErr] ()

Do some basic sanity checking on the Shelley genesis file.

mkShelleyGlobals ∷ ShelleyGenesis era → EpochInfo Identity → Natural → Globals

data SnapShots era

Snapshots of the stake distribution.

Constructors

SnapShots

Fields

_pstakeMark ∷ !(SnapShot era)

_pstakeSet ∷ !(SnapShot era)

_pstakeGo ∷ !(SnapShot era)

_feeSS ∷ !Coin

data SnapShot era

Snapshot of the stake distribution.

Constructors

SnapShot

Fields

_stake ∷ !(Stake era)

_delegations ∷ !(Map (Credential 'Staking era) (KeyHash 'StakePool (Crypto era)))

_poolParams ∷ !(Map (KeyHash 'StakePool (Crypto era)) (PoolParams era))

newtype Stake era

Type of stake as map from hash key to coins associated.

Constructors

Stake

Fields

unStake ∷ Map (Credential 'Staking era) Coin

data NonMyopic era

newtype WitHashes era

Constructors

WitHashes

Fields

unWitHashes ∷ Set (KeyHash 'Witness (Crypto era))

data LedgerState era

The state associated with a Ledger.

Constructors

LedgerState

Fields

_utxoState ∷ !(UTxOState era)

The current unspent transaction outputs.

_delegationState ∷ !(DPState era)

The current delegation state

data NewEpochState era

New Epoch state and environment

Constructors

NewEpochState

Fields

nesEL ∷ !EpochNo

Last epoch

nesBprev ∷ !(BlocksMade era)

Blocks made before current epoch

nesBcur ∷ !(BlocksMade era)

Blocks made in current epoch

nesEs ∷ !(EpochState era)

Epoch state before current

nesRu ∷ !(StrictMaybe (RewardUpdate era))

Possible reward update

nesPd ∷ !(PoolDistr (Crypto era))

Stake distribution within the stake pool

data UTxOState era

Constructors

UTxOState

Fields

_utxo ∷ !(UTxO era)

_deposited ∷ !Coin

_fees ∷ !Coin

_ppups ∷ !(PPUPState era)

data PPUPState era

Constructors

PPUPState

Fields

proposals ∷ !(ProposedPPUpdates era)

futureProposals ∷ !(ProposedPPUpdates era)

data EpochState era

Constructors

EpochState

Fields

esAccountState ∷ !AccountState

esSnapshots ∷ !(SnapShots era)

esLState ∷ !(LedgerState era)

esPrevPp ∷ !(PParams era)

esPp ∷ !(PParams era)

esNonMyopic ∷ !(NonMyopic era)

This field, esNonMyopic, does not appear in the formal spec and is not a part of the protocol. It is only used for providing data to the stake pool ranking calculation getNonMyopicMemberRewards. See https://hydra.iohk.io/job/Cardano/cardano-ledger-specs/specs.pool-ranking/latest/download-by-type/doc-pdf/pool-ranking

data AccountState

Constructors

AccountState

Fields

_treasury ∷ !Coin

_reserves ∷ !Coin

data RewardUpdate era

Constructors

RewardUpdate

Fields

deltaT ∷ !Coin

deltaR ∷ !DeltaCoin

rs ∷ !(Map (Credential 'Staking era) Coin)

deltaF ∷ !DeltaCoin

nonMyopic ∷ !(NonMyopic era)

data DPState era

The state associated with the current stake delegation.

Constructors

DPState

Fields

_dstate ∷ !(DState era)

_pstate ∷ !(PState era)

data PState era

Current state of staking pools and their certificate counters.

Constructors

PState

Fields

_pParams ∷ !(Map (KeyHash 'StakePool (Crypto era)) (PoolParams era))

The pool parameters.

_fPParams ∷ !(Map (KeyHash 'StakePool (Crypto era)) (PoolParams era))

The future pool parameters.

_retiring ∷ !(Map (KeyHash 'StakePool (Crypto era)) EpochNo)

A map of retiring stake pools to the epoch when they retire.

data DState era

State of staking pool delegations and rewards

Constructors

DState

Fields

_rewards ∷ !(RewardAccounts era)

The active reward accounts.

_delegations ∷ !(Map (Credential 'Staking era) (KeyHash 'StakePool (Crypto era)))

The current delegations.

_ptrs ∷ !(Bimap Ptr (Credential 'Staking era))

The pointed to hash keys.

_fGenDelegs ∷ !(Map (FutureGenDeleg (Crypto era)) (GenDelegPair (Crypto era)))

future genesis key delegations

_genDelegs ∷ !(GenDelegs (Crypto era))

Genesis key delegations

_irwd ∷ !(InstantaneousRewards era)

Instantaneous Rewards

data InstantaneousRewards era

Constructors

InstantaneousRewards

Fields

iRReserves ∷ !(Map (Credential 'Staking era) Coin)

iRTreasury ∷ !(Map (Credential 'Staking era) Coin)

type KeyPairs era = [(KeyPair 'Payment era, KeyPair 'Staking era)]

Representation of a list of pairs of key pairs, e.g., pay and stake keys

data PPUPEnv era

Constructors

PPUPEnv SlotNo (PParams era) (GenDelegs (Crypto era))

data PPUP era

data UtxoEnv era

Constructors

UtxoEnv SlotNo (PParams era) (Map (KeyHash 'StakePool (Crypto era)) (PoolParams era)) (GenDelegs (Crypto era))

data UTXO era

data UTXOW era

data POOLREAP era

data PoolEnv era

Constructors

PoolEnv SlotNo (PParams era)

data POOL (era ∷ Type)

data NEWEPOCH era

calculatePoolDistr ∷ SnapShot era → PoolDistr (Crypto era)

data TICK era

data DelegEnv

Constructors

DelegEnv

Fields

slotNo ∷ SlotNo

ptr_ ∷ Ptr

acnt_ ∷ AccountState

data DELEG era

data DelplEnv era

Constructors

DelplEnv

Fields

delplSlotNo ∷ SlotNo

delPlPtr ∷ Ptr

delPlPp ∷ PParams era

delPlAcnt ∷ AccountState

data DELPL era

data DelegsEnv era

Constructors

DelegsEnv

Fields

delegsSlotNo ∷ SlotNo

delegsIx ∷ Ix

delegspp ∷ PParams era

delegsTx ∷ Tx era

delegsAccount ∷ AccountState

data DELEGS era

data LedgerEnv era

Constructors

LedgerEnv

Fields

ledgerSlotNo ∷ SlotNo

ledgerIx ∷ Ix

ledgerPp ∷ PParams era

ledgerAccount ∷ AccountState

data LEDGER era

data LedgersEnv era

Constructors

LedgersEnv

Fields

ledgersSlotNo ∷ SlotNo

ledgersPp ∷ PParams era

ledgersAccount ∷ AccountState

data LEDGERS era

newtype LaxBlock era

Constructors

LaxBlock (Block era)

data Block era

Constructors

Block' !(BHeader (Crypto era)) !(TxSeq era) ByteString

Bundled Patterns

pattern Block ∷ Era era ⇒ BHeader (Crypto era) → TxSeq era → Block era

data BHBody crypto

Constructors

BHBody

Fields

bheaderBlockNo ∷ !BlockNo

block number

bheaderSlotNo ∷ !SlotNo

block slot

bheaderPrev ∷ !(PrevHash crypto)

Hash of the previous block header

bheaderVk ∷ !(VKey 'BlockIssuer crypto)

verification key of block issuer

bheaderVrfVk ∷ !(VerKeyVRF crypto)

VRF verification key for block issuer

bheaderEta ∷ !(CertifiedVRF crypto Nonce)

block nonce

bheaderL ∷ !(CertifiedVRF crypto Natural)

leader election value

bsize ∷ !Natural

Size of the block body

bhash ∷ !(HashBBody crypto)

Hash of block body

bheaderOCert ∷ !(OCert crypto)

operational certificate

bprotver ∷ !ProtVer

protocol version

data PrevHash crypto

The previous hash of a block

Constructors

GenesisHash

BlockHash !(HashHeader crypto)

data BHeader crypto where

Bundled Patterns

pattern BHeader ∷ Crypto crypto ⇒ BHBody crypto → SignedKES crypto (BHBody crypto) → BHeader crypto

newtype HashBBody crypto

Hash of block body

Constructors

UnsafeHashBBody

Fields

unHashBody ∷ Hash crypto EraIndependentBlockBody

newtype HashHeader crypto

The hash of a Block Header

Constructors

HashHeader

Fields

unHashHeader ∷ Hash crypto (BHeader crypto)

bhHash ∷ ∀ crypto. Crypto crypto ⇒ BHeader crypto → HashHeader crypto

Hash a given block header

bbHash ∷ ∀ era. Era era ⇒ TxSeq era → HashBBody (Crypto era)

Hash a given block body

bHeaderSize ∷ ∀ crypto. Crypto crypto ⇒ BHeader crypto → Int

bheader ∷ Era era ⇒ Block era → BHeader (Crypto era)

bbody ∷ Era era ⇒ Block era → TxSeq era

bhbody ∷ Crypto crypto ⇒ BHeader crypto → BHBody crypto

data PrtlSeqFailure crypto

Constructors

WrongSlotIntervalPrtclSeq

Fields

SlotNo

Last slot number.

SlotNo

Current slot number.

WrongBlockNoPrtclSeq

Fields

(WithOrigin (LastAppliedBlock crypto))

Last applied block.

BlockNo

Current block number.

WrongBlockSequencePrtclSeq

Fields

(PrevHash crypto)

Last applied hash

(PrevHash crypto)

Current block's previous hash

data PrtclPredicateFailure crypto

Constructors

OverlayFailure (PredicateFailure (OVERLAY crypto))

UpdnFailure (PredicateFailure (UPDN crypto))

data PrtclEnv crypto

Constructors

PrtclEnv UnitInterval (PoolDistr crypto) (GenDelegs crypto) Nonce

data PrtclState crypto

Constructors

PrtclState

Fields

!(Map (KeyHash 'BlockIssuer crypto) Word64)

Operation Certificate counters

!Nonce

Evolving nonce

!Nonce

Candidate nonce

prtlSeqChecks ∷ (MonadError (PrtlSeqFailure crypto) m, Crypto crypto) ⇒ WithOrigin (LastAppliedBlock crypto) → BHeader crypto → m ()

data ChainState era

Constructors

ChainState

Fields

chainNes ∷ NewEpochState era

chainOCertIssue ∷ Map (KeyHash 'BlockIssuer (Crypto era)) Word64

chainEpochNonce ∷ Nonce

chainEvolvingNonce ∷ Nonce

chainCandidateNonce ∷ Nonce

chainPrevEpochNonce ∷ Nonce

chainLastAppliedBlock ∷ WithOrigin (LastAppliedBlock (Crypto era))

data CHAIN era

initialShelleyState ∷ WithOrigin (LastAppliedBlock (Crypto era)) → EpochNo → UTxO era → Coin → Map (KeyHash 'Genesis (Crypto era)) (GenDelegPair (Crypto era)) → PParams era → Nonce → ChainState era

Creates a valid initial chain state