Shelley.Spec.Ledger.API.Wallet

Documentation

getNonMyopicMemberRewards ∷ Era era ⇒ Globals → ShelleyState era → Set (Either Coin (Credential Staking era)) → Map (Either Coin (Credential Staking era)) (Map (KeyHash StakePool era) Coin)

Calculate the Non-Myopic Pool Member Rewards for a set of credentials. For each given credential, this function returns a map from each stake pool (identified by the key hash of the pool operator) to the non-myopic pool member reward for that stake pool.

This is not based on any snapshot, but uses the current ledger state.

getUTxO ∷ ShelleyState era → UTxO era

Get the full UTxO.

getFilteredUTxO ∷ ShelleyState era → Set (Addr era) → UTxO era

Get the UTxO filtered by address.

getLeaderSchedule ∷ (Era era, Signable (VRF (Crypto era)) Seed) ⇒ Globals → ShelleyState era → ChainDepState era → KeyHash StakePool era → SignKeyVRF era → PParams era → Set SlotNo

Get the (private) leader schedule for this epoch.

Given a private VRF key, returns the set of slots in which this node is eligible to lead.

getTotalStake ∷ Globals → ShelleyState era → Coin

Calculate the current total stake.

poolsByTotalStakeFraction ∷ Era era ⇒ Globals → ShelleyState era → PoolDistr era

Get pool sizes, but in terms of total stake

The stake distribution uses active stake (so that the leader schedule is not affected by undelegated stake), but the wallet wants to display pool saturation for rewards purposes. For that, it needs the fraction of total stake.

This is not based on any snapshot, but uses the current ledger state.