Shelley.Spec.Ledger.Address

Documentation

mkVKeyRwdAcnt ∷ Crypto crypto ⇒ KeyPair Staking crypto → RewardAcnt crypto

mkRwdAcnt ∷ Credential Staking crypto → RewardAcnt crypto

scriptsToAddr ∷ Crypto crypto ⇒ (MultiSig crypto, MultiSig crypto) → Addr crypto

Create a base address from a pair of multi-sig scripts (pay and stake)

scriptToCred ∷ Crypto crypto ⇒ MultiSig crypto → Credential kr crypto

Convert a given multi-sig script to a credential by hashing it and wrapping into the Credential data type.

TODO nc what is the role of this credential?

toAddr ∷ Crypto crypto ⇒ (KeyPair Payment crypto, KeyPair Staking crypto) → Addr crypto

toCred ∷ Crypto crypto ⇒ KeyPair kr crypto → Credential kr crypto

serialiseAddr ∷ Crypto crypto ⇒ Addr crypto → ByteString

Serialise an address to the external format.

deserialiseAddr ∷ Crypto crypto ⇒ ByteString → Maybe (Addr crypto)

Deserialise an address from the external format. This will fail if the input data is not in the right format (or if there is trailing data).

data Addr crypto

An address for UTxO.

Constructors

Addr !(PaymentCredential crypto) !(StakeReference crypto)

AddrBootstrap !Address

getAddr ∷ ∀ crypto. Crypto crypto ⇒ Get (Addr crypto)

getKeyHash ∷ Crypto crypto ⇒ Get (Credential kr crypto)

getPtr ∷ Get Ptr

getScriptHash ∷ Crypto crypto ⇒ Get (Credential kr crypto)

getVariableLengthNat ∷ Get Natural

putAddr ∷ ∀ crypto. Crypto crypto ⇒ Addr crypto → Put

putCredential ∷ Credential kr crypto → Put

putPtr ∷ Ptr → Put

putVariableLengthNat ∷ Natural → Put

natToWord7s ∷ Natural → [Word7]

word7sToNat ∷ [Word7] → Natural

newtype Word7

Constructors

Word7 Word8