Shelley.Spec.Ledger.BaseTypes

Documentation

type FixedPoint = Digits34

(==>) ∷ Bool → Bool → Bool infix 1

(⭒) ∷ Nonce → Nonce → Nonce

Evolve the nonce

data Network

Constructors

Testnet

Mainnet

networkToWord8 ∷ Network → Word8

word8ToNetwork ∷ Word8 → Maybe Network

data Nonce

Evolving nonce type.

Constructors

Nonce !(Hash Blake2b_256 Nonce)

NeutralNonce
Identity element

newtype Seed

Seed to the verifiable random function.

We do not expose the constructor to Seed. Instead, a Seed should be created using mkSeed for a VRF calculation.

Constructors

Seed (Hash Blake2b_256 Seed)

data UnitInterval

Type to represent a value in the unit interval [0; 1]

fpPrecision ∷ FixedPoint

interval0 ∷ UnitInterval

intervalValue ∷ UnitInterval → Ratio Word64

Get rational value of UnitInterval type

unitIntervalToRational ∷ UnitInterval → Rational

unitIntervalFromRational ∷ Rational → UnitInterval

invalidKey ∷ Word → Decoder s a

mkNonceFromOutputVRF ∷ OutputVRF v → Nonce

Make a nonce from the VRF output bytes

mkNonceFromNumber ∷ Word64 → Nonce

Make a nonce from a number.

mkUnitInterval ∷ Ratio Word64 → Maybe UnitInterval

Return a UnitInterval type if r is in [0; 1].

truncateUnitInterval ∷ Ratio Word64 → UnitInterval

Convert a rational to a UnitInterval by ignoring its integer part.

data StrictMaybe a

Strict Maybe.

TODO move to cardano-prelude

Constructors

SNothing

SJust !a

strictMaybeToMaybe ∷ StrictMaybe a → Maybe a

maybeToStrictMaybe ∷ Maybe a → StrictMaybe a

fromSMaybe ∷ a → StrictMaybe a → a

data Url

urlToText ∷ Url → Text

textToUrl ∷ Text → Maybe Url

data DnsName

dnsToText ∷ DnsName → Text

textToDns ∷ Text → Maybe DnsName

newtype Port

Constructors

Port

Fields

portToWord16 ∷ Word16

data ActiveSlotCoeff

mkActiveSlotCoeff ∷ UnitInterval → ActiveSlotCoeff

activeSlotVal ∷ ActiveSlotCoeff → UnitInterval

activeSlotLog ∷ ActiveSlotCoeff → FixedPoint

STS Base

data Globals

Constructors

Globals

Fields

epochInfo ∷ !(EpochInfo Identity)

slotsPerKESPeriod ∷ !Word64

stabilityWindow ∷ !Word64

The window size in which our chosen chain growth property guarantees at least k blocks. From the paper "Ouroboros praos: An adaptively-secure, semi-synchronous proof-of-stake protocol". The stabilityWindow constant is used in a number of places; for example, protocol updates must be submitted at least twice this many slots before an epoch boundary.

randomnessStabilisationWindow ∷ !Word64

Number of slots before the end of the epoch at which we stop updating the candidate nonce for the next epoch.

securityParameter ∷ !Word64

Maximum number of blocks we are allowed to roll back

maxKESEvo ∷ !Word64

Maximum number of KES iterations

quorum ∷ !Word64

Quorum for update system votes and MIR certificates

maxMajorPV ∷ !Natural

All blocks invalid after this protocol version

maxLovelaceSupply ∷ !Word64

Maximum number of lovelace in the system

activeSlotCoeff ∷ !ActiveSlotCoeff

Active Slot Coefficient, named f in "Ouroboros Praos: An adaptively-secure, semi-synchronous proof-of-stake protocol"

networkId ∷ !Network

The network ID

type ShelleyBase = ReaderT Globals Identity