Shelley.Spec.Ledger.ByronTranslation

Documentation

mkInitialShelleyLedgerView ∷ ∀ era. ShelleyGenesis era → LedgerView era

We construct a LedgerView using the Shelley genesis config in the same way as translateToShelleyLedgerState.

translateToShelleyLedgerState ∷ ∀ era. (Era era, ADDRHASH (Crypto era) ~ Blake2b_224) ⇒ ShelleyGenesis era → EpochNo → ChainValidationState → ShelleyState era

Exported for testing purposes

translateCompactTxOutByronToShelley ∷ CompactTxOut → TxOut era

translateTxIdByronToShelley ∷ (Era era, ADDRHASH (Crypto era) ~ Blake2b_224) ⇒ TxId → TxId era

We use the same hashing algorithm so we can unwrap and rewrap the bytes. We don't care about the type that is hashed, which will differ going from Byron to Shelley, we just use the hashes as IDs.