Shelley.Spec.Ledger.Core

Documentation

class Relation m where

Minimal complete definition

singleton, dom, range, (◁), (⋪), (▷), (⋫), (∪), (⨃), size

Associated Types

type Domain m ∷ Type

type Range m ∷ Type

Methods

singleton ∷ Domain m → Range m → m

dom ∷ Ord (Domain m) ⇒ m → Set (Domain m)

Domain

range ∷ Ord (Range m) ⇒ m → Set (Range m)

Range

(◁) ∷ Ord (Domain m) ⇒ Set (Domain m) → m → m

Domain restriction

Unicode: 25c1

(<|) ∷ Ord (Domain m) ⇒ Set (Domain m) → m → m

Domain restriction

Unicode: 25c1

(⋪) ∷ Ord (Domain m) ⇒ Set (Domain m) → m → m

Domain exclusion

Unicode: 22ea

(</|) ∷ Ord (Domain m) ⇒ Set (Domain m) → m → m

Domain exclusion

Unicode: 22ea

(▷) ∷ Ord (Range m) ⇒ m → Set (Range m) → m

Range restriction

Unicode: 25b7

(|>) ∷ Ord (Range m) ⇒ m → Set (Range m) → m

Range restriction

Unicode: 25b7

(⋫) ∷ Ord (Range m) ⇒ m → Set (Range m) → m

Range exclusion

Unicode: 22eb

(|/>) ∷ Ord (Range m) ⇒ m → Set (Range m) → m

Range exclusion

Unicode: 22eb

(∪) ∷ (Ord (Domain m), Ord (Range m)) ⇒ m → m → m

Union

(⨃) ∷ (Ord (Domain m), Ord (Range m)) ⇒ m → m → m

Union Override Right

size ∷ Integral n ⇒ m → n

Size of the relation

haskey ∷ Ord (Domain m) ⇒ Domain m → m → Bool

Is this key in the Domain, Instances should overide this default with something more efficient

addpair ∷ (Ord (Domain m), Ord (Range m)) ⇒ Domain m → Range m → m → m

Insert (key,value) pair into the Relation. Instances should overide this default with something more efficient

removekey ∷ Ord (Domain m) ⇒ Domain m → m → m

Remove a key (and its associted value at that key) from the Relation. Instances should overide this default with something more efficient

(⊆) ∷ (Foldable f, Foldable g, Ord a) ⇒ f a → g a → Bool

Inclusion among foldables.

Unicode: 2286

(∪+) ∷ (Ord a, Num b) ⇒ Map a b → Map a b → Map a b

Union override plus is (AB)∪(BA)∪{k|->v1+v2 | k|->v1 : A / k|->v2 : B} The library function Map.unionWith is more general, it allows any type for b as long as (+) :: b -> b -> b

(∈) ∷ (Eq a, Foldable f) ⇒ a → f a → Bool

Alias for elem.

Unicode: 2208

(∉) ∷ (Eq a, Foldable f) ⇒ a → f a → Bool infixl 4

Alias for not elem.

Unicode: 2209

(∩) ∷ Ord a ⇒ Set a → Set a → Set a