Shelley.Spec.Ledger.Credential

Documentation

data Credential (kr ∷ KeyRole) era

Script hash or key hash for a payment or a staking object.

Constructors

ScriptHashObj !(ScriptHash era)

KeyHashObj !(KeyHash kr era)

newtype GenesisCredential era

Constructors

GenesisCredential

Fields

unGenesisCredential ∷ KeyHash Genesis era

type Ix = Natural

type PaymentCredential era = Credential Payment era

data Ptr

Pointer to a slot, transaction index and index in certificate list.

Constructors

Ptr !SlotNo !Ix !Ix

type StakeCredential era = Credential Staking era

data StakeReference era

Constructors

StakeRefBase !(StakeCredential era)

StakeRefPtr !Ptr

StakeRefNull