Shelley.Spec.Ledger.Delegation.Certificates

Documentation

data DCert era

A heavyweight certificate.

Constructors

DCertDeleg !(DelegCert era)

DCertPool !(PoolCert era)

DCertGenesis !(GenesisDelegCert era)

DCertMir !(MIRCert era)

data DelegCert era

Constructors

RegKey !(StakeCredential era)
A stake key registration certificate.

DeRegKey !(StakeCredential era)
A stake key deregistration certificate.

Delegate !(Delegation era)
A stake delegation certificate.

data PoolCert era

Constructors

RegPool !(PoolParams era)
A stake pool registration certificate.

RetirePool !(KeyHash StakePool era) !EpochNo
A stake pool retirement certificate.

data GenesisDelegCert era

Genesis key delegation certificate

Constructors

GenesisDelegCert !(KeyHash Genesis era) !(KeyHash GenesisDelegate era) !(Hash era (VerKeyVRF era))

data MIRCert era

Move instantaneous rewards certificate

Constructors

MIRCert

Fields

mirPot ∷ MIRPot

mirRewards ∷ Map (Credential Staking era) Coin

newtype StakeCreds era

Constructors

StakeCreds

Fields

unStakeCreds ∷ Map (Credential Staking era) SlotNo

newtype PoolDistr era

Constructors

PoolDistr

Fields

unPoolDistr ∷ Map (KeyHash StakePool era) (IndividualPoolStake era)

data IndividualPoolStake era

Constructors

IndividualPoolStake

Fields

individualPoolStake ∷ !Rational

individualPoolStakeVrf ∷ !(Hash era (VerKeyVRF era))

delegCWitness ∷ DelegCert era → Credential Staking era

Determine the certificate author

poolCWitness ∷ PoolCert era → Credential StakePool era

genesisCWitness ∷ GenesisDelegCert era → KeyHash Genesis era

isRegKey ∷ DCert era → Bool

Check for RegKey constructor

isDeRegKey ∷ DCert era → Bool

Check for DeRegKey constructor

isDelegation ∷ DCert era → Bool

Check for Delegation constructor

isGenesisDelegation ∷ DCert era → Bool

Check for GenesisDelegate constructor

isRegPool ∷ DCert era → Bool

Check for RegPool constructor

isRetirePool ∷ DCert era → Bool

Check for RetirePool constructor

isInstantaneousRewards ∷ DCert era → Bool

isReservesMIRCert ∷ DCert era → Bool

isTreasuryMIRCert ∷ DCert era → Bool

requiresVKeyWitness ∷ DCert era → Bool

Returns True for delegation certificates that require at least one witness, and False otherwise. It is mainly used to ensure that calling a variant of cwitness is safe.