Shelley.Spec.Ledger.EpochBoundary

Description

This modules implements the necessary functions for the changes that can happen at epoch boundaries.

Documentation

newtype Stake era

Type of stake as map from hash key to coins associated.

Constructors

Stake

Fields

unStake ∷ Map (Credential Staking era) Coin

newtype BlocksMade era

Blocks made

Constructors

BlocksMade

Fields

unBlocksMade ∷ Map (KeyHash StakePool era) Natural

data SnapShot era

Snapshot of the stake distribution.

Constructors

SnapShot

Fields

_stake ∷ !(Stake era)

_delegations ∷ !(Map (Credential Staking era) (KeyHash StakePool era))

_poolParams ∷ !(Map (KeyHash StakePool era) (PoolParams era))

data SnapShots era

Snapshots of the stake distribution.

Constructors

SnapShots

Fields

_pstakeMark ∷ !(SnapShot era)

_pstakeSet ∷ !(SnapShot era)

_pstakeGo ∷ !(SnapShot era)

_feeSS ∷ !Coin

emptySnapShot ∷ SnapShot era

emptySnapShots ∷ SnapShots era

aggregateUtxoCoinByCredential ∷ Era era ⇒ Map Ptr (Credential Staking era) → UTxO era → Map (Credential Staking era) Coin → Map (Credential Staking era) Coin

Sum up all the Coin for each staking Credential

poolStake ∷ KeyHash StakePool era → Map (Credential Staking era) (KeyHash StakePool era) → Stake era → Stake era

Get stake of one pool

obligation ∷ PParams era → Map (Credential Staking era) Coin → Map (KeyHash StakePool era) (PoolParams era) → Coin

Calculate total possible refunds.

maxPool ∷ PParams era → Coin → Rational → Rational → Coin

Calculate maximal pool reward