Shelley.Spec.Ledger.Keys

Documentation

data KeyRole

The role of a key.

Note that a role is not _fixed_, nor is it unique. In particular, keys may variously be used as witnesses, and so in many case we will change the role of a key to the Witness role.

It is also perfectly allowable for a key to be used in many roles; there is nothing prohibiting somebody using the same underlying key as their payment and staking key, as well as the key for their stake pool. So these roles are more intended for two purposes:

To make explicit how we are using a key in the specifications

To provide a guide to downstream implementors, for whom the profusion of keys may be confusing.

Constructors

Genesis

GenesisDelegate

Payment

Staking

StakePool

BlockIssuer

Witness

class HasKeyRole (a ∷ KeyRole → Type → Type) where

Minimal complete definition

Nothing

Methods

coerceKeyRole ∷ a r crypto → a r' crypto

General coercion of key roles.

The presence of this function is mostly to help the user realise where they are converting key roles.

default coerceKeyRole ∷ Coercible (a r crypto) (a r' crypto) ⇒ a r crypto → a r' crypto

asWitness ∷ HasKeyRole a ⇒ a r crypto → a 'Witness crypto

Use a key as a witness.

This is the most common coercion between key roles, because most keys can be used as witnesses to some types of transaction. As such, we provide an explicit coercion for it.

DSIGN

type DSignable c = Signable (DSIGN c)

newtype VKey (kd ∷ KeyRole) crypto

Discriminated verification key

We wrap the basic VerKeyDSIGN in order to add the key role.

Constructors

VKey

Fields

unVKey ∷ VerKeyDSIGN (DSIGN crypto)

data KeyPair (kd ∷ KeyRole) crypto

Pair of signing key and verification key, with a usage role.

Constructors

KeyPair

Fields

vKey ∷ !(VKey kd crypto)

sKey ∷ !(SignKeyDSIGN (DSIGN crypto))

signedDSIGN ∷ (Crypto crypto, Signable (DSIGN crypto) a) ⇒ SignKeyDSIGN (DSIGN crypto) → a → SignedDSIGN crypto a

Produce a digital signature

verifySignedDSIGN ∷ (Crypto crypto, Signable (DSIGN crypto) a) ⇒ VKey kd crypto → a → SignedDSIGN crypto a → Bool

Verify a digital signature

Key hashes

newtype KeyHash (discriminator ∷ KeyRole) crypto

Discriminated hash of public Key

Constructors

KeyHash (Hash (ADDRHASH crypto) (VerKeyDSIGN (DSIGN crypto)))

hashKey ∷ Crypto crypto ⇒ VKey kd crypto → KeyHash kd crypto

Hash a given public key

Genesis delegations

data GenDelegPair crypto

Constructors

GenDelegPair

Fields

genDelegKeyHash ∷ !(KeyHash 'GenesisDelegate crypto)

genDelegVrfHash ∷ !(Hash crypto (VerKeyVRF crypto))

newtype GenDelegs crypto

Constructors

GenDelegs

Fields

unGenDelegs ∷ Map (KeyHash 'Genesis crypto) (GenDelegPair crypto)

newtype GKeys crypto

Constructors

GKeys

Fields

unGKeys ∷ Set (VKey 'Genesis crypto)

KES

type KESignable c = Signable (KES c)

VRF

type VRFSignable c = Signable (VRF c)

Re-exports from cardano-crypto-class

decodeSignedDSIGN ∷ DSIGNAlgorithm v ⇒ Decoder s (SignedDSIGN v a)

encodeSignedDSIGN ∷ DSIGNAlgorithm v ⇒ SignedDSIGN v a → Encoding

hashWithSerialiser ∷ HashAlgorithm h ⇒ (a → Encoding) → a → Hash h a

A variation on hashWith, but specially for CBOR encodings.

decodeSignedKES ∷ KESAlgorithm v ⇒ Decoder s (SignedKES v a)

decodeVerKeyKES ∷ KESAlgorithm v ⇒ Decoder s (VerKeyKES v)

encodeSignedKES ∷ KESAlgorithm v ⇒ SignedKES v a → Encoding

encodeVerKeyKES ∷ KESAlgorithm v ⇒ VerKeyKES v → Encoding

signedKES ∷ (KESAlgorithm v, Signable v a) ⇒ ContextKES v → Period → a → SignKeyKES v → SignedKES v a

updateKES

Arguments

∷ KESAlgorithm v

⇒ ContextKES v

→ SignKeyKES v

→ Period
The current period for the key, not the target period.

→ Maybe (SignKeyKES v)

Update the KES signature key to the next period, given the current period.

It returns Nothing if the cannot be evolved any further.

The precondition (to get a Just result) is that the current KES period of the input key is not the last period. The given period must be the current KES period of the input key (not the next or target).

The postcondition is that in case a key is returned, its current KES period is incremented by one compared to before.

Note that you must track the current period separately, and to skip to a later period requires repeated use of this function, since it only increments one period at once.

verifyKES

Arguments

∷ (KESAlgorithm v, Signable v a, HasCallStack)

⇒ ContextKES v

→ VerKeyKES v

→ Period
The current period for the key

→ a

→ SigKES v

→ Either String ()

verifySignedKES ∷ (KESAlgorithm v, Signable v a) ⇒ ContextKES v → VerKeyKES v → Period → a → SignedKES v a → Either String ()

decodeVerKeyVRF ∷ VRFAlgorithm v ⇒ Decoder s (VerKeyVRF v)

encodeVerKeyVRF ∷ VRFAlgorithm v ⇒ VerKeyVRF v → Encoding

hashVerKeyVRF ∷ (VRFAlgorithm v, HashAlgorithm h) ⇒ VerKeyVRF v → Hash h (VerKeyVRF v)

verifyVRF ∷ (VRFAlgorithm v, HasCallStack, Signable v a) ⇒ ContextVRF v → VerKeyVRF v → a → (OutputVRF v, CertVRF v) → Bool

Re-parametrised types over crypto

type CertifiedVRF c = CertifiedVRF (VRF c)

type Hash c = Hash (HASH c)

type SignedDSIGN c = SignedDSIGN (DSIGN c)

type SignKeyDSIGN c = SignKeyDSIGN (DSIGN c)

type SignedKES c = SignedKES (KES c)

type SignKeyKES c = SignKeyKES (KES c)

type SignKeyVRF c = SignKeyVRF (VRF c)

type VerKeyKES c = VerKeyKES (KES c)

type VerKeyVRF c = VerKeyVRF (VRF c)