Shelley.Spec.Ledger.LedgerState

Description

This module implements the operation rules for treating UTxO transactions (Tx) as state transformations on a ledger state (LedgerState), as specified in A Simplified Formal Specification of a UTxO Ledger.

Documentation

data AccountState

Constructors

AccountState

Fields

_treasury ∷ !Coin

_reserves ∷ !Coin

data DPState era

The state associated with the current stake delegation.

Constructors

DPState

Fields

_dstate ∷ !(DState era)

_pstate ∷ !(PState era)

data DState era

State of staking pool delegations and rewards

Constructors

DState

Fields

_rewards ∷ !(RewardAccounts era)

The active reward accounts.

_delegations ∷ !(Map (Credential Staking era) (KeyHash StakePool era))

The current delegations.

_ptrs ∷ !(Bimap Ptr (Credential Staking era))

The pointed to hash keys.

_fGenDelegs ∷ !(Map (FutureGenDeleg era) (GenDelegPair era))

future genesis key delegations

_genDelegs ∷ !(GenDelegs era)

Genesis key delegations

_irwd ∷ !(InstantaneousRewards era)

Instantaneous Rewards

data EpochState era

Constructors

EpochState

Fields

esAccountState ∷ !AccountState

esSnapshots ∷ !(SnapShots era)

esLState ∷ !(LedgerState era)

esPrevPp ∷ !(PParams era)

esPp ∷ !(PParams era)

esNonMyopic ∷ !(NonMyopic era)

data FutureGenDeleg era

Constructors

FutureGenDeleg

Fields

fGenDelegSlot ∷ !SlotNo

fGenDelegGenKeyHash ∷ !(KeyHash Genesis era)

data InstantaneousRewards era

Constructors

InstantaneousRewards

Fields

iRReserves ∷ !(Map (Credential Staking era) Coin)

iRTreasury ∷ !(Map (Credential Staking era) Coin)

type Ix = Natural

type KeyPairs era = [(KeyPair Payment era, KeyPair Staking era)]

Representation of a list of pairs of key pairs, e.g., pay and stake keys

data LedgerState era

The state associated with a Ledger.

Constructors

LedgerState

Fields

_utxoState ∷ !(UTxOState era)

The current unspent transaction outputs.

_delegationState ∷ !(DPState era)

The current delegation state

data PPUPState era

Constructors

PPUPState

Fields

proposals ∷ !(ProposedPPUpdates era)

futureProposals ∷ !(ProposedPPUpdates era)

data PState era

Current state of staking pools and their certificate counters.

Constructors

PState

Fields

_pParams ∷ !(Map (KeyHash StakePool era) (PoolParams era))

The pool parameters.

_fPParams ∷ !(Map (KeyHash StakePool era) (PoolParams era))

The future pool parameters.

_retiring ∷ !(Map (KeyHash StakePool era) EpochNo)

A map of retiring stake pools to the epoch when they retire.

type RewardAccounts era = Map (Credential Staking era) Coin

data RewardUpdate era

Constructors

RewardUpdate

Fields

deltaT ∷ !Coin

deltaR ∷ !Coin

rs ∷ !(Map (Credential Staking era) Coin)

deltaF ∷ !Coin

nonMyopic ∷ !(NonMyopic era)

data UTxOState era

Constructors

UTxOState

Fields

_utxo ∷ !(UTxO era)

_deposited ∷ !Coin

_fees ∷ !Coin

_ppups ∷ !(PPUPState era)

depositPoolChange ∷ Era era ⇒ LedgerState era → PParams era → TxBody era → Coin

Calculate the change to the deposit pool for a given transaction.

emptyAccount ∷ AccountState

emptyDPState ∷ DPState era

emptyDState ∷ DState era

emptyEpochState ∷ EpochState era

emptyInstantaneousRewards ∷ InstantaneousRewards era

emptyLedgerState ∷ LedgerState era

emptyPPUPState ∷ PPUPState era

emptyPState ∷ PState era

emptyRewardUpdate ∷ RewardUpdate era

emptyUTxOState ∷ UTxOState era

pvCanFollow ∷ ProtVer → StrictMaybe ProtVer → Bool

reapRewards ∷ RewardAccounts era → RewardAccounts era → RewardAccounts era

totalInstantaneousReservesRewards ∷ InstantaneousRewards era → Coin

updatePpup ∷ UTxOState era → PParams era → UTxOState era

Update the protocol parameter updates by clearing out the proposals and making the future proposals become the new proposals, provided the new proposals can follow (otherwise reset them).

state transitions

emptyDelegation ∷ DPState era

Genesis State

genesisState ∷ Map (KeyHash Genesis era) (GenDelegPair era) → UTxO era → LedgerState era

Creates the ledger state for an empty ledger which contains the specified transaction outputs.

Validation

newtype WitHashes era

Constructors

WitHashes

Fields

unWitHashes ∷ Set (KeyHash Witness era)

nullWitHashes ∷ WitHashes era → Bool

Check if a set of witness hashes is empty.

diffWitHashes ∷ WitHashes era → WitHashes era → WitHashes era

Extract the difference between two sets of witness hashes.

minfee ∷ PParams era → Tx era → Coin

Minimum fee calculation

minfeeBound ∷ ∀ era. Era era ⇒ PParams era → Tx era → Coin

Minimum fee bound using txsizeBound

txsize ∷ Tx era → Integer

Implementation of abstract transaction size

txsizeBound ∷ ∀ era. Era era ⇒ Tx era → Integer

Convenience Function to bound the txsize function. | It can be helpful for coin selection.

produced ∷ Era era ⇒ PParams era → Map (KeyHash StakePool era) (PoolParams era) → TxBody era → Coin

Compute the lovelace which are created by the transaction

consumed ∷ Era era ⇒ PParams era → UTxO era → TxBody era → Coin

Compute the lovelace which are destroyed by the transaction

verifiedWits ∷ (Era era, DSignable era (Hash era (TxBody era))) ⇒ Tx era → Either [VKey Witness era] ()

Given a ledger state, determine if the UTxO witnesses in a given transaction are correct.

witsVKeyNeeded ∷ ∀ era. Era era ⇒ UTxO era → Tx era → GenDelegs era → WitHashes era

Collect the set of hashes of keys that needs to sign a given transaction. This set consists of the txin owners, certificate authors, and withdrawal reward accounts.

witsFromWitnessSet ∷ Era era ⇒ WitnessSet era → WitHashes era

Extract the witness hashes from the Witness set.

DelegationState

keyRefunds ∷ Era era ⇒ PParams era → TxBody era → Coin

Compute the key deregistration refunds in a transaction

Epoch boundary

stakeDistr ∷ ∀ era. Era era ⇒ UTxO era → DState era → PState era → SnapShot era

applyRUpd ∷ RewardUpdate era → EpochState era → EpochState era

Apply a reward update

createRUpd ∷ EpochSize → BlocksMade era → EpochState era → Coin → ShelleyBase (RewardUpdate era)

Create a reward update

data NewEpochState era

New Epoch state and environment

Constructors

NewEpochState

Fields

nesEL ∷ !EpochNo

Last epoch

nesBprev ∷ !(BlocksMade era)

Blocks made before current epoch

nesBcur ∷ !(BlocksMade era)

Blocks made in current epoch

nesEs ∷ !(EpochState era)

Epoch state before current

nesRu ∷ !(StrictMaybe (RewardUpdate era))

Possible reward update

nesPd ∷ !(PoolDistr era)

Stake distribution within the stake pool

getGKeys ∷ NewEpochState era → Set (KeyHash Genesis era)

updateNES ∷ NewEpochState era → BlocksMade era → LedgerState era → NewEpochState era

Update new epoch state

circulation ∷ EpochState era → Coin → Coin

Calculate the current circulation

This is used in the rewards calculation, and for API endpoints for pool ranking.