Shelley.Spec.Ledger.OverlaySchedule

Overlay schedule

isOverlaySlot ∷ SlotNo → UnitInterval → SlotNo → Bool

classifyOverlaySlot ∷ SlotNo → Set (KeyHash 'Genesis crypto) → UnitInterval → ActiveSlotCoeff → SlotNo → OBftSlot crypto

lookupInOverlaySchedule ∷ SlotNo → Set (KeyHash 'Genesis crypto) → UnitInterval → ActiveSlotCoeff → SlotNo → Maybe (OBftSlot crypto)

OBftSlot

data OBftSlot crypto

Constructors

NonActiveSlot

ActiveSlot !(KeyHash 'Genesis crypto)

Testing

overlaySlots ∷ SlotNo → UnitInterval → EpochSize → [SlotNo]

Return the list of overlaySlots for a given epoch. Note that this linear in the size of the epoch, and should probably only be used for testing. If something more performant is needed, we could probably use [start + floor(xd) | x <- [0 .. (spe -1)], floor(xd) < spe] but we would need to make sure that this is equivalent.