Shelley.Spec.Ledger.PParams

Description

This module contains just the type of protocol parameters.

Documentation

data PParams' f

Protocol parameters.

We use the HKD type family so that the protocol parameters type and the type for the updates to the protocol parameters can share records fields. The protocol parameters will have type PParams' Identity, and the updates will have type PParams' StrictMaybe, though Identity will be hidden from use.

For example:

  myParameters =
    PParams
      { _minfeeA = 0,
        _minfeeB = 0,
        ...
      }

  myUpdate =
    PParamsUpdate
      { _minfeeA = SNothing,
        _minfeeB = SJust 42,
        ...
      }

Constructors

PParams

Fields

_minfeeA ∷ !(HKD f Natural)

The linear factor for the minimum fee calculation

_minfeeB ∷ !(HKD f Natural)

The constant factor for the minimum fee calculation

_maxBBSize ∷ !(HKD f Natural)

Maximal block body size

_maxTxSize ∷ !(HKD f Natural)

Maximal transaction size

_maxBHSize ∷ !(HKD f Natural)

Maximal block header size

_keyDeposit ∷ !(HKD f Coin)

The amount of a key registration deposit

_keyMinRefund ∷ !(HKD f UnitInterval)

The minimum percent refund guarantee

_keyDecayRate ∷ !(HKD f Rational)

The deposit decay rate

_poolDeposit ∷ !(HKD f Coin)

The amount of a pool registration deposit

_poolMinRefund ∷ !(HKD f UnitInterval)

The minimum percent pool refund

_poolDecayRate ∷ !(HKD f Rational)

Decay rate for pool deposits

_eMax ∷ !(HKD f EpochNo)

epoch bound on pool retirement

_nOpt ∷ !(HKD f Natural)

Desired number of pools

_a0 ∷ !(HKD f Rational)

Pool influence

_rho ∷ !(HKD f UnitInterval)

Treasury expansion

_tau ∷ !(HKD f UnitInterval)

Monetary expansion

_d ∷ !(HKD f UnitInterval)

Decentralization parameter

_extraEntropy ∷ !(HKD f Nonce)

Extra entropy

_protocolVersion ∷ !(HKD f ProtVer)

Protocol version

_minUTxOValue ∷ !(HKD f Natural)

Minimum UTxO value

type PParams = PParams' Identity

emptyPParams ∷ PParams

Returns a basic "empty" PParams' structure with all zero values.

data ProtVer

Constructors

ProtVer !Natural !Natural

data PPUpdateEnv crypto

Constructors

PPUpdateEnv SlotNo (GenDelegs crypto)

newtype ProposedPPUpdates crypto

Update operation for protocol parameters structure @PParams

Constructors

ProposedPPUpdates (Map (KeyHash Genesis crypto) PParamsUpdate)

emptyPPPUpdates ∷ ProposedPPUpdates crypto

type PParamsUpdate = PParams' StrictMaybe

data Update crypto

Update Proposal

Constructors

Update !(ProposedPPUpdates crypto) !EpochNo

updatePParams ∷ PParams → PParamsUpdate → PParams