Shelley.Spec.Ledger.STS.Overlay

Documentation

data OVERLAY era

type family PredicateFailure a = (b ∷ Type) | b → a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data OverlayEnv era

Constructors

OverlayEnv UnitInterval (PoolDistr era) (GenDelegs era) Nonce

data OverlayPredicateFailure era

Constructors

VRFKeyUnknown !(KeyHash StakePool era)

VRFKeyWrongVRFKey !(KeyHash StakePool era) !(Hash era (VerKeyVRF era)) !(Hash era (VerKeyVRF era))

VRFKeyBadNonce !Nonce !SlotNo !Nonce !(CertifiedVRF (VRF (Crypto era)) Nonce)

VRFKeyBadLeaderValue !Nonce !SlotNo !Nonce !(CertifiedVRF (VRF (Crypto era)) Nonce)

VRFLeaderValueTooBig !(OutputVRF (VRF (Crypto era))) !Rational !ActiveSlotCoeff

NotActiveSlotOVERLAY !SlotNo

WrongGenesisColdKeyOVERLAY !(KeyHash BlockIssuer era) !(KeyHash GenesisDelegate era)

WrongGenesisVRFKeyOVERLAY !(KeyHash BlockIssuer era) !(Hash era (VerKeyVRF era)) !(Hash era (VerKeyVRF era))

UnknownGenesisKeyOVERLAY !(KeyHash Genesis era)

OcertFailure (PredicateFailure (OCERT era))