Shelley.Spec.Ledger.STS.Prtcl

Documentation

data PRTCL crypto

type family State a

Type of the state which the system transitions between.

data PrtclEnv crypto

Constructors

PrtclEnv UnitInterval (PoolDistr crypto) (GenDelegs crypto) Nonce

data PrtclState crypto

Constructors

PrtclState

Fields

!(Map (KeyHash 'BlockIssuer crypto) Word64)

Operation Certificate counters

!Nonce

Evolving nonce

!Nonce

Candidate nonce

data PrtclPredicateFailure crypto

Constructors

OverlayFailure (PredicateFailure (OVERLAY crypto))

UpdnFailure (PredicateFailure (UPDN crypto))

type family PredicateFailure a = (b ∷ Type) | b → a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data PrtlSeqFailure crypto

Constructors

WrongSlotIntervalPrtclSeq

Fields

SlotNo

Last slot number.

SlotNo

Current slot number.

WrongBlockNoPrtclSeq

Fields

(WithOrigin (LastAppliedBlock crypto))

Last applied block.

BlockNo

Current block number.

WrongBlockSequencePrtclSeq

Fields

(PrevHash crypto)

Last applied hash

(PrevHash crypto)

Current block's previous hash

prtlSeqChecks ∷ (MonadError (PrtlSeqFailure crypto) m, Crypto crypto) ⇒ WithOrigin (LastAppliedBlock crypto) → BHeader crypto → m ()