Shelley.Spec.Ledger.STS.Tick

Documentation

data TICK era

type family State a

Type of the state which the system transitions between.

data TickPredicateFailure era

Constructors

NewEpochFailure (PredicateFailure (NEWEPOCH era))

RupdFailure (PredicateFailure (RUPD era))

type family PredicateFailure a = (b ∷ Type) | b → a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

adoptGenesisDelegs ∷ EpochState era → SlotNo → EpochState era

data TICKF era

data TickfPredicateFailure era

Constructors

TickfNewEpochFailure (PredicateFailure (NEWEPOCH era))