Shelley.Spec.Ledger.Scripts

Documentation

data MultiSig era where

Bundled Patterns

pattern RequireAllOf ∷ Era era ⇒ [MultiSig era] → MultiSig era

pattern RequireAnyOf ∷ Era era ⇒ [MultiSig era] → MultiSig era

pattern RequireSignature ∷ Era era ⇒ KeyHash Witness era → MultiSig era

pattern RequireMOf ∷ Era era ⇒ Int → [MultiSig era] → MultiSig era

data Script era

Constructors

MultiSigScript (MultiSig era)

newtype ScriptHash era

Constructors

ScriptHash (Hash (ADDRHASH (Crypto era)) (Script era))

getKeyCombination ∷ Era era ⇒ MultiSig era → [KeyHash Witness era]

Get one possible combination of keys for multi signature script

getKeyCombinations ∷ Era era ⇒ MultiSig era → [[KeyHash Witness era]]

Get all valid combinations of keys for given multi signature. This is mainly useful for testing.

hashAnyScript ∷ Era era ⇒ Script era → ScriptHash era

hashMultiSigScript ∷ Era era ⇒ MultiSig era → ScriptHash era

Hashes native multi-signature script.