Shelley.Spec.Ledger.Slot

Documentation

newtype SlotNo

The 0-based index for the Ourboros time slot.

Constructors

SlotNo

Fields

unSlotNo ∷ Word64

newtype Duration

Constructors

Duration Word64

(-*) ∷ SlotNo → SlotNo → Duration

(+*) ∷ SlotNo → Duration → SlotNo

(*-) ∷ SlotNo → Duration → SlotNo

Subtract a duration from a slot

newtype EpochNo

An epoch, i.e. the number of the epoch.

Constructors

EpochNo

Fields

unEpochNo ∷ Word64

newtype EpochSize

Constructors

EpochSize

Fields

unEpochSize ∷ Word64

data EpochInfo (m ∷ Type → Type)

Information about epochs

Epochs may have different sizes at different times during the lifetime of the blockchain. This information is encapsulated by EpochInfo; it is parameterized over a monad m because the information about how long each epoch is may depend on information derived from the blockchain itself, and hence requires access to state.

The other functions provide some derived information from epoch sizes. In the default implementation all of these functions query and update an internal cache maintaining cumulative epoch sizes; for that reason, all of these functions live in a monad m.

slotByronToShelley ∷ Slot → SlotNo

Convert SlotNo data from Byron to Shelley.

slotShelleyToByron ∷ SlotNo → Slot

newtype BlockNo

The 0-based index of the block in the blockchain. BlockNo is <= SlotNo and is only equal at slot N if there is a block for every slot where N <= SlotNo.

Constructors

BlockNo

Fields

unBlockNo ∷ Word64

epochInfoEpoch ∷ EpochInfo Identity → SlotNo → ShelleyBase EpochNo

epochInfoFirst ∷ EpochInfo Identity → EpochNo → ShelleyBase SlotNo

epochInfoSize ∷ EpochInfo Identity → EpochNo → ShelleyBase EpochSize