Shelley.Spec.Ledger.Tx

Documentation

data Tx era

A fully formed transaction.

Constructors

Tx' !(TxBody era) !(WitnessSet era) !(StrictMaybe MetaData) ByteString

Bundled Patterns

pattern Tx ∷ TxBodyConstraints era ⇒ TxBody era → WitnessSet era → StrictMaybe MetaData → Tx era

newtype TxBody era

Constructors

TxBodyY (MemoBytes (TxBodyX era))

Bundled Patterns

pattern TxBody ∷ ProperTo era ⇒ Set (TxIn era) → StrictSeq (TxOut era) → StrictSeq (DCert era) → Wdrl era → Coin → SlotNo → StrictMaybe (Update era) → StrictMaybe (MetaDataHash era) → TxBody era
Pattern for use by external users

data TxOut era

The output of a UTxO.

Constructors

TxOutCompact !(CompactAddr era) !(CompactForm (Value era))

Bundled Patterns

pattern TxOut ∷ ShelleyBased era ⇒ Addr era → Value era → TxOut era

data TxIn era

The input of a UTxO.

Constructors

TxInCompact !(TxId era) !Word64

Bundled Patterns

pattern TxIn ∷ Era era ⇒ TxId era → Natural → TxIn era

newtype TxId era

A unique ID of a transaction, which is computable from the transaction.

Constructors

TxId

Fields

_unTxId ∷ Hash (Crypto era) EraIndependentTxBody

decodeWits ∷ ∀ era s. (TxBodyConstraints era, AnnotatedData (Script era), ValidateScript era) ⇒ Decoder s (Annotator (WitnessSet era))

segwitTx ∷ TxBodyConstraints era ⇒ Annotator (TxBody era) → Annotator (WitnessSet era) → Maybe (Annotator MetaData) → Annotator (Tx era)

type WitnessSet = WitnessSetHKD Identity

data WitnessSetHKD f era where

Bundled Patterns

pattern WitnessSet ∷ (Era era, AnnotatedData (Script era)) ⇒ Set (WitVKey 'Witness era) → Map (ScriptHash era) (Script era) → Set (BootstrapWitness era) → WitnessSet era

data WitVKey kr era where

Proof/Witness that a transaction is authorized by the given key holder.

Bundled Patterns

pattern WitVKey ∷ (Typeable kr, Era era) ⇒ VKey kr (Crypto era) → SignedDSIGN (Crypto era) (Hash (Crypto era) EraIndependentTxBody) → WitVKey kr era

class (Era era, ToCBOR (Script era)) ⇒ ValidateScript era where

Typeclass for multis-signature script data types. Allows for script validation and hashing.

Methods

validateScript ∷ Script era → Tx era → Bool

hashScript ∷ Script era → ScriptHash era

txwitsScript ∷ TxBodyConstraints era ⇒ Tx era → Map (ScriptHash era) (Script era)

Multi-signature script witness accessor function for Transactions

extractKeyHashWitnessSet ∷ ∀ (r ∷ KeyRole) era. [Credential r era] → Set (KeyHash 'Witness (Crypto era))

getKeyCombinations ∷ Era era ⇒ MultiSig era → [[KeyHash 'Witness (Crypto era)]]

Get all valid combinations of keys for given multi signature. This is mainly useful for testing.

getKeyCombination ∷ Era era ⇒ MultiSig era → [KeyHash 'Witness (Crypto era)]

Get one possible combination of keys for multi signature script

addrWits' ∷ WitnessSetHKD f era → HKD f (Set (WitVKey 'Witness era))

evalNativeMultiSigScript ∷ Era era ⇒ MultiSig era → Set (KeyHash 'Witness (Crypto era)) → Bool

Script evaluator for native multi-signature scheme. vhks is the set of key hashes that signed the transaction to be validated.

hashMultiSigScript ∷ Era era ⇒ MultiSig era → ScriptHash era

Hashes native multi-signature script.

validateNativeMultiSigScript ∷ TxBodyConstraints era ⇒ MultiSig era → Tx era → Bool

Script validator for native multi-signature scheme.