Shelley.Spec.Ledger.TxBody

Documentation

data DCert era

A heavyweight certificate.

Constructors

DCertDeleg !(DelegCert era)

DCertPool !(PoolCert era)

DCertGenesis !(GenesisDelegCert era)

DCertMir !(MIRCert era)

data DelegCert era

Constructors

RegKey !(StakeCredential era)
A stake key registration certificate.

DeRegKey !(StakeCredential era)
A stake key deregistration certificate.

Delegate !(Delegation era)
A stake delegation certificate.

data Delegation era

The delegation of one stake key to another.

Constructors

Delegation

Fields

_delegator ∷ !(StakeCredential era)

_delegatee ∷ !(KeyHash StakePool era)

data GenesisDelegCert era

Genesis key delegation certificate

Constructors

GenesisDelegCert !(KeyHash Genesis era) !(KeyHash GenesisDelegate era) !(Hash era (VerKeyVRF era))

type Ix = Natural

data MIRCert era

Move instantaneous rewards certificate

Constructors

MIRCert

Fields

mirPot ∷ MIRPot

mirRewards ∷ Map (Credential Staking era) Coin

data MIRPot

Constructors

ReservesMIR

TreasuryMIR

data PoolCert era

Constructors

RegPool !(PoolParams era)
A stake pool registration certificate.

RetirePool !(KeyHash StakePool era) !EpochNo
A stake pool retirement certificate.

data PoolMetaData

Constructors

PoolMetaData

Fields

_poolMDUrl ∷ !Url

_poolMDHash ∷ !ByteString

data PoolParams era

A stake pool.

Constructors

PoolParams

Fields

_poolPubKey ∷ !(KeyHash StakePool era)

_poolVrf ∷ !(Hash era (VerKeyVRF era))

_poolPledge ∷ !Coin

_poolCost ∷ !Coin

_poolMargin ∷ !UnitInterval

_poolRAcnt ∷ !(RewardAcnt era)

_poolOwners ∷ !(Set (KeyHash Staking era))

_poolRelays ∷ !(StrictSeq StakePoolRelay)

_poolMD ∷ !(StrictMaybe PoolMetaData)

data Ptr

Pointer to a slot, transaction index and index in certificate list.

Constructors

Ptr !SlotNo !Ix !Ix

data RewardAcnt era

An account based address for rewards

Constructors

RewardAcnt

Fields

getRwdNetwork ∷ !Network

getRwdCred ∷ !(Credential Staking era)

newtype StakeCreds era

Constructors

StakeCreds

Fields

unStakeCreds ∷ Map (Credential Staking era) SlotNo

data StakePoolRelay

Constructors

SingleHostAddr !(StrictMaybe Port) !(StrictMaybe IPv4) !(StrictMaybe IPv6)
One or both of IPv4 & IPv6

SingleHostName !(StrictMaybe Port) !DnsName
An A or AAAA DNS record

MultiHostName !DnsName
A SRV DNS record

data TxBody era where

A raw transaction

Bundled Patterns

pattern TxBody ∷ Era era ⇒ Set (TxIn era) → StrictSeq (TxOut era) → StrictSeq (DCert era) → Wdrl era → Coin → SlotNo → StrictMaybe (Update era) → StrictMaybe (MetaDataHash era) → TxBody era

newtype TxId era

A unique ID of a transaction, which is computable from the transaction.

Constructors

TxId

Fields

_unTxId ∷ Hash era (TxBody era)

data TxIn era where

The input of a UTxO.

Bundled Patterns

pattern TxIn ∷ Era era ⇒ TxId era → Natural → TxIn era

pattern TxInCompact ∷ !(TxId era) → !Word64 → TxIn era

data TxOut era

The output of a UTxO.

Constructors

TxOutCompact !ShortByteString !Word64

Bundled Patterns

pattern TxOut ∷ Era era ⇒ Addr era → Coin → TxOut era

data Url

newtype Wdrl era

Constructors

Wdrl

Fields

unWdrl ∷ Map (RewardAcnt era) Coin

data WitVKey era kr where

Proof/Witness that a transaction is authorized by the given key holder.

Bundled Patterns

pattern WitVKey ∷ (Typeable kr, Era era) ⇒ VKey kr era → SignedDSIGN era (Hash era (TxBody era)) → WitVKey era kr

witKeyHash ∷ WitVKey era kr → KeyHash Witness era

data SizeOfPoolOwners

The size of the _poolOwners Set. Only used to compute size of encoded PoolParams.

Constructors

SizeOfPoolOwners

data SizeOfPoolRelays

The size of the _poolRelays Set. Only used to compute size of encoded PoolParams.

Constructors

SizeOfPoolRelays