Shelley.Spec.Ledger.TxData

Documentation

data DCert crypto

A heavyweight certificate.

Constructors

DCertDeleg !(DelegCert crypto)

DCertPool !(PoolCert crypto)

DCertGenesis !(GenesisDelegCert crypto)

DCertMir !(MIRCert crypto)

data DelegCert crypto

Constructors

RegKey !(StakeCredential crypto)
A stake key registration certificate.

DeRegKey !(StakeCredential crypto)
A stake key deregistration certificate.

Delegate !(Delegation crypto)
A stake delegation certificate.

data Delegation crypto

The delegation of one stake key to another.

Constructors

Delegation

Fields

_delegator ∷ !(StakeCredential crypto)

_delegatee ∷ !(KeyHash StakePool crypto)

data GenesisDelegCert crypto

Genesis key delegation certificate

Constructors

GenesisDelegCert !(KeyHash Genesis crypto) !(KeyHash GenesisDelegate crypto)

type Ix = Natural

newtype MIRCert crypto

Move instantaneous rewards certificate

Constructors

MIRCert (Map (Credential Staking crypto) Coin)

data PoolCert crypto

Constructors

RegPool !(PoolParams crypto)
A stake pool registration certificate.

RetirePool !(KeyHash StakePool crypto) !EpochNo
A stake pool retirement certificate.

data PoolMetaData

Constructors

PoolMetaData

Fields

_poolMDUrl ∷ !Url

_poolMDHash ∷ !ByteString

data PoolParams crypto

A stake pool.

Constructors

PoolParams

Fields

_poolPubKey ∷ !(KeyHash StakePool crypto)

_poolVrf ∷ !(Hash crypto (VerKeyVRF crypto))

_poolPledge ∷ !Coin

_poolCost ∷ !Coin

_poolMargin ∷ !UnitInterval

_poolRAcnt ∷ !(RewardAcnt crypto)

_poolOwners ∷ !(Set (KeyHash Staking crypto))

_poolRelays ∷ !(StrictSeq StakePoolRelay)

_poolMD ∷ !(StrictMaybe PoolMetaData)

data Ptr

Pointer to a slot, transaction index and index in certificate list.

Constructors

Ptr !SlotNo !Ix !Ix

newtype RewardAcnt crypto

An account based address for rewards

Constructors

RewardAcnt

Fields

getRwdCred ∷ Credential Staking crypto

newtype StakeCreds crypto

Constructors

StakeCreds (Map (Credential Staking crypto) SlotNo)

newtype StakePools crypto

Constructors

StakePools

Fields

unStakePools ∷ Map (KeyHash StakePool crypto) SlotNo

data StakePoolRelay

Constructors

SingleHostAddr !(StrictMaybe Port) !(StrictMaybe IPv4) !(StrictMaybe IPv6)
One or both of IPv4 & IPv6

SingleHostName !(StrictMaybe Port) !DnsName
An A or AAAA DNS record

MultiHostName !(StrictMaybe Port) !DnsName
A SRV DNS record

data TxBody crypto where

A raw transaction

Bundled Patterns

pattern TxBody ∷ Crypto crypto ⇒ Set (TxIn crypto) → StrictSeq (TxOut crypto) → StrictSeq (DCert crypto) → Wdrl crypto → Coin → SlotNo → StrictMaybe (Update crypto) → StrictMaybe (MetaDataHash crypto) → TxBody crypto

newtype TxId crypto

A unique ID of a transaction, which is computable from the transaction.

Constructors

TxId

Fields

_TxId ∷ Hash crypto (TxBody crypto)

data TxIn crypto

The input of a UTxO.

Constructors

TxIn !(TxId crypto) !Natural

data TxOut crypto

The output of a UTxO.

Constructors

TxOut !(Addr crypto) !Coin

data Url

newtype Wdrl crypto

Constructors

Wdrl

Fields

unWdrl ∷ Map (RewardAcnt crypto) Coin

data WitVKey crypto where

Proof/Witness that a transaction is authorized by the given key holder.

Bundled Patterns

pattern WitVKey ∷ Crypto crypto ⇒ VKey Witness crypto → SignedDSIGN crypto (Hash crypto (TxBody crypto)) → WitVKey crypto

witKeyHash ∷ ∀ crypto. Crypto crypto ⇒ WitVKey crypto → KeyHash Witness crypto

addStakeCreds ∷ Credential Staking crypto → SlotNo → StakeCreds crypto → StakeCreds crypto

data SizeOfPoolOwners

The size of the _poolOwners Set. Only used to compute size of encoded PoolParams.

Constructors

SizeOfPoolOwners

data SizeOfPoolRelays

The size of the _poolRelays Set. Only used to compute size of encoded PoolParams.

Constructors

SizeOfPoolRelays