Shelley.Spec.Ledger.UTxO

Description

This module defines the types and functions for a simple UTxO Ledger as specified in A Simplified Formal Specification of a UTxO Ledger.

Primitives

newtype UTxO era

The unspent transaction outputs.

Constructors

UTxO

Fields

unUTxO ∷ Map (TxIn era) (TxOut era)

Functions

txid ∷ Era era ⇒ TxBody era → TxId era

Compute the id of a transaction.

txins ∷ Era era ⇒ TxBody era → Set (TxIn era)

Compute the UTxO inputs of a transaction.

txinLookup ∷ TxIn era → UTxO era → Maybe (TxOut era)

Lookup a txin for a given UTxO collection

txouts ∷ Era era ⇒ TxBody era → UTxO era

Compute the transaction outputs of a transaction.

txup ∷ Era era ⇒ Tx era → Maybe (Update era)

balance ∷ UTxO era → Coin

Determine the total balance contained in the UTxO.

totalDeposits ∷ PParams era → Map (KeyHash StakePool era) (PoolParams era) → [DCert era] → Coin

Determine the total deposit amount needed. The block may (legitimately) contain multiple registration certificates for the same pool, where the first will be treated as a registration and any subsequent ones as re-registration. As such, we must only take a deposit for the first such registration.

Note that this is not an issue for key registrations since subsequent registration certificates would be invalid.

makeWitnessVKey ∷ ∀ era kr. (Era era, DSignable era (Hash era (TxBody era))) ⇒ Hash era (TxBody era) → KeyPair kr era → WitVKey era Witness

Create a witness for transaction

makeWitnessesVKey ∷ ∀ era kr. (Era era, DSignable era (Hash era (TxBody era))) ⇒ Hash era (TxBody era) → [KeyPair kr era] → Set (WitVKey era Witness)

Create witnesses for transaction

makeWitnessesFromScriptKeys ∷ (Era era, DSignable era (Hash era (TxBody era))) ⇒ Hash era (TxBody era) → Map (KeyHash kr era) (KeyPair kr era) → Set (KeyHash kr era) → Set (WitVKey era Witness)

From a list of key pairs and a set of key hashes required for a multi-sig scripts, return the set of required keys.

verifyWitVKey ∷ (Typeable kr, Era era, DSignable era (Hash era (TxBody era))) ⇒ Hash era (TxBody era) → WitVKey era kr → Bool

Verify a transaction body witness

scriptsNeeded ∷ Era era ⇒ UTxO era → Tx era → Set (ScriptHash era)

Computes the set of script hashes required to unlock the transcation inputs and the withdrawals.

txinsScript ∷ Era era ⇒ Set (TxIn era) → UTxO era → Set (TxIn era)

Compute the subset of inputs of the set txInps for which each input is locked by a script in the UTxO u.