Shelley.Spec.Ledger.Validation

Documentation

data ValidationError

Validation errors represent the failures of a transaction to be valid for a given ledger state.

Constructors

BadInputs
The transaction inputs are not valid.

Expired SlotNo SlotNo
The transaction has expired

RetirementCertExpired SlotNo SlotNo
Pool Retirement Certificate expired

FeeTooSmall Coin Coin
The transaction fee is too small

ValueNotConserved Coin Coin
Value is not conserved

IncorrectRewards
Unknown reward account

InvalidWitness
One of the transaction witnesses is invalid.

MissingWitnesses
The transaction does not have the required witnesses.

InputSetEmpty
Missing Replay Attack Protection, at least one input must be spent.

StakeKeyAlreadyRegistered
A stake key cannot be registered again.

StakeKeyNotRegistered
A stake key must be registered to be used or deregistered.

StakeDelegationImpossible
The stake key to which is delegated is not known.

StakePoolNotRegisteredOnKey
Stake pool not registered for key, cannot be retired.

UnknownValidationError
Other error (for conversion from STS errors that have no correspondance here)

data Validity

The validity of a transaction, where an invalid transaction is represented by list of errors.

Constructors

Valid

Invalid [ValidationError]