Shelley.Spec.Ledger.API.Protocol

Description

Integration between the Shelley ledger and its corresponding (Transitional Praos) protocol.

In particular, this code supports extracting the components of the ledger state needed for protocol execution, both now and in a 2k-slot window.

Documentation

data PrtclEnv era

data LedgerView era

Data required by the Transitional Praos protocol from the Shelley ledger.

Constructors

LedgerView

Fields

lvProtParams ∷ PParams era

lvPoolDistr ∷ PoolDistr era

lvGenDelegs ∷ GenDelegs era

currentLedgerView ∷ ShelleyState era → LedgerView era

Alias of view for export

Time Travel (or the anachronistic ledger view)

The ledger needs to expose access to the LedgerView for a window of slots around the current tip of the chain. We call this period the stability window, and it corresponds to the number of slots needed to "guarantee" the presence of k blocks (where k is the security parameter). This functionality allows the protocol layer to validate headers without downloading corresponding blocks.

The ability to travel backwards in time is obviously always possible by keeping a record of past ledger states (or, more conservatively, ledger views). We do not therefore deal explicitly with it in this module, though see later for discussion on when snapshots should be taken.

In order to achieve forward time travel, we need a few things: - Transition rules which process the body of a block should not have any effect on the LedgerView during the stability window after they are received. This property should be guaranteed by the design of the ledger. - The effect of transition rules which process the header of a block should be predictable for the stability window.

We make the following claim:

A future ledger view (within the stability window) is equal to the application of the TICK rule at the target slot to the curernt ledger state.

futureLedgerView ∷ ∀ era m. (Era era, MonadError (FutureLedgerViewError era) m) ⇒ Globals → ShelleyState era → SlotNo → m (LedgerView era)

Anachronistic ledger view

Given a slot within the future stability window from our current slot (the slot corresponding to the passed-in ShelleyState), return a LedgerView appropriate to that slot.

Chain state operations

The chain state is an amalgam of the protocol state and the ticked nonce.

data ChainDepState c

Constructors

ChainDepState

Fields

csProtocol ∷ !(PrtclState c)

csTickn ∷ !TicknState

csLabNonce ∷ !Nonce

Nonce constructed from the hash of the last applied block header.

newtype ChainTransitionError era

Constructors

ChainTransitionError [PredicateFailure (PRTCL era)]

tickChainDepState

Arguments

∷ Era era

⇒ Globals

→ LedgerView era

→ Bool
Are we in a new epoch?

→ ChainDepState era

→ ChainDepState era

Tick the chain state to a new epoch.

updateChainDepState ∷ ∀ era m. (Era era, MonadError (ChainTransitionError era) m, Signable (DSIGN (Crypto era)) (OCertSignable era), Signable (KES (Crypto era)) (BHBody era), Signable (VRF (Crypto era)) Seed) ⇒ Globals → LedgerView era → BHeader era → ChainDepState era → m (ChainDepState era)

Update the chain state based upon a new block header.

This also updates the last applied block hash.

reupdateChainDepState ∷ ∀ era. (Era era, Signable (DSIGN (Crypto era)) (OCertSignable era), Signable (KES (Crypto era)) (BHBody era), Signable (VRF (Crypto era)) Seed) ⇒ Globals → LedgerView era → BHeader era → ChainDepState era → ChainDepState era

Re-update the chain state based upon a new block header.

This function does no validation of whether the header is internally valid or consistent with the chain it is being applied to; the caller must ensure that this is valid through having previously applied it.