Shelley.Spec.Ledger.API.Protocol

Description

Integration between the Shelley ledger and its corresponding (Transitional Praos) protocol.

In particular, this code supports extracting the components of the ledger state needed for protocol execution, both now and in a 2k-slot window.

Documentation

data PrtclEnv crypto

mkPrtclEnv

Arguments

∷ LedgerView crypto

→ Bool
New epoch marker. This should be true iff this execution of the PRTCL rule is being run on the first block in a new epoch.

→ PrtclEnv crypto

Construct a protocol environment from the ledger view, along with the current slot and a marker indicating whether this is the first block in a new epoch.

data LedgerView crypto

Data required by the Transitional Praos protocol from the Shelley ledger.

Constructors

LedgerView

Fields

lvProtParams ∷ PParams

lvOverlaySched ∷ Map SlotNo (OBftSlot crypto)

lvPoolDistr ∷ PoolDistr crypto

lvGenDelegs ∷ GenDelegs crypto

currentLedgerView ∷ ShelleyState crypto → LedgerView crypto

Alias of view for export

Time Travel (or the anachronistic ledger view)

The ledger needs to expose access to the LedgerView for a window of slots around the current tip of the chain. We call this period the stability window, and it corresponds to the number of slots needed to "guarantee" the presence of k blocks (where k is the security parameter). This functionality allows the protocol layer to validate headers without downloading corresponding blocks.

The ability to travel backwards in time is obviously always possible by keeping a record of past ledger states (or, more conservatively, ledger views). We do not therefore deal explicitly with it in this module, though see later for discussion on when snapshots should be taken.

In order to achieve forward time travel, we need a few things: - Transition rules which process the body of a block should not have any effect on the LedgerView during the stability window after they are received. This property should be guaranteed by the design of the ledger. - The effect of transition rules which process the header of a block should be predictable for the stability window.

We make the following claim:

A future ledger view (within the stability window) is equal to the application of the TICK rule at the target slot to the curernt ledger state.

futureLedgerView ∷ ∀ crypto m. (Crypto crypto, MonadError (FutureLedgerViewError crypto) m) ⇒ Globals → ShelleyState crypto → SlotNo → m (LedgerView crypto)

Anachronistic ledger view

Given a slot within the future stability window from our current slot (the slot corresponding to the passed-in ShelleyState), return a LedgerView appropriate to that slot.