Shelley.Spec.Ledger.API.Types

Documentation

hashVerKeyVRF ∷ (VRFAlgorithm v, HashAlgorithm h) ⇒ VerKeyVRF v → Hash h (VerKeyVRF v)

newtype Coin

The amount of value held by a transaction output.

Constructors

Coin

Fields

unCoin ∷ Integer

type VerKeyVRF e = VerKeyVRF (VRF (Crypto e))

type SignKeyVRF e = SignKeyVRF (VRF (Crypto e))

type CertifiedVRF e = CertifiedVRF (VRF (Crypto e))

type VerKeyKES e = VerKeyKES (KES (Crypto e))

type SignKeyKES e = SignKeyKES (KES (Crypto e))

type SignKeyDSIGN e = SignKeyDSIGN (DSIGN (Crypto e))

type SignedDSIGN e = SignedDSIGN (DSIGN (Crypto e))

newtype GenDelegs era

Constructors

GenDelegs

Fields

unGenDelegs ∷ Map (KeyHash Genesis era) (GenDelegPair era)

data GenDelegPair era

Constructors

GenDelegPair

Fields

genDelegKeyHash ∷ !(KeyHash GenesisDelegate era)

genDelegVrfHash ∷ !(Hash era (VerKeyVRF era))

newtype KeyHash (discriminator ∷ KeyRole) era

Discriminated hash of public Key

Constructors

KeyHash (Hash (ADDRHASH (Crypto era)) (VerKeyDSIGN (DSIGN (Crypto era))))

data KeyPair (kd ∷ KeyRole) era

Pair of signing key and verification key, with a usage role.

Constructors

KeyPair

Fields

vKey ∷ !(VKey kd era)

sKey ∷ !(SignKeyDSIGN (DSIGN (Crypto era)))

newtype VKey (kd ∷ KeyRole) era

Discriminated verification key

We wrap the basic VerKeyDSIGN in order to add the key role.

Constructors

VKey

Fields

unVKey ∷ VerKeyDSIGN (DSIGN (Crypto era))

data KeyRole

The role of a key.

Note that a role is not _fixed_, nor is it unique. In particular, keys may variously be used as witnesses, and so in many case we will change the role of a key to the Witness role.

It is also perfectly allowable for a key to be used in many roles; there is nothing prohibiting somebody using the same underlying key as their payment and staking key, as well as the key for their stake pool. So these roles are more intended for two purposes:

To make explicit how we are using a key in the specifications

To provide a guide to downstream implementors, for whom the profusion of keys may be confusing.

Constructors

Genesis

GenesisDelegate

Payment

Staking

StakePool

BlockIssuer

Witness

data Network

Constructors

Testnet

Mainnet

newtype Port

Constructors

Port

Fields

portToWord16 ∷ Word16

data StrictMaybe a

Strict Maybe.

TODO move to cardano-prelude

Constructors

SNothing

SJust !a

data Script era

Constructors

MultiSigScript (MultiSig era)

newtype ScriptHash era

Constructors

ScriptHash (Hash (ADDRHASH (Crypto era)) (Script era))

data MultiSig era where

Bundled Patterns

pattern RequireAllOf ∷ Era era ⇒ [MultiSig era] → MultiSig era

pattern RequireAnyOf ∷ Era era ⇒ [MultiSig era] → MultiSig era

pattern RequireSignature ∷ Era era ⇒ KeyHash Witness era → MultiSig era

pattern RequireMOf ∷ Era era ⇒ Int → [MultiSig era] → MultiSig era

data OBftSlot era

Constructors

NonActiveSlot

ActiveSlot !(KeyHash Genesis era)

newtype ProposedPPUpdates era

Update operation for protocol parameters structure @PParams

Constructors

ProposedPPUpdates (Map (KeyHash Genesis era) (PParamsUpdate era))

data Update era

Update Proposal

Constructors

Update !(ProposedPPUpdates era) !EpochNo

type PParams era = PParams' Identity era

data PParams' f era

Protocol parameters.

We use the HKD type family so that the protocol parameters type and the type for the updates to the protocol parameters can share records fields. The protocol parameters will have type PParams' Identity, and the updates will have type PParams' StrictMaybe, though Identity will be hidden from use.

For example:

  myParameters =
    PParams
      { _minfeeA = 0,
        _minfeeB = 0,
        ...
      }

  myUpdate =
    PParamsUpdate
      { _minfeeA = SNothing,
        _minfeeB = SJust 42,
        ...
      }

Constructors

PParams

Fields

_minfeeA ∷ !(HKD f Natural)

The linear factor for the minimum fee calculation

_minfeeB ∷ !(HKD f Natural)

The constant factor for the minimum fee calculation

_maxBBSize ∷ !(HKD f Natural)

Maximal block body size

_maxTxSize ∷ !(HKD f Natural)

Maximal transaction size

_maxBHSize ∷ !(HKD f Natural)

Maximal block header size

_keyDeposit ∷ !(HKD f Coin)

The amount of a key registration deposit

_poolDeposit ∷ !(HKD f Coin)

The amount of a pool registration deposit

_eMax ∷ !(HKD f EpochNo)

epoch bound on pool retirement

_nOpt ∷ !(HKD f Natural)

Desired number of pools

_a0 ∷ !(HKD f Rational)

Pool influence

_rho ∷ !(HKD f UnitInterval)

Monetary expansion

_tau ∷ !(HKD f UnitInterval)

Treasury expansion

_d ∷ !(HKD f UnitInterval)

Decentralization parameter

_extraEntropy ∷ !(HKD f Nonce)

Extra entropy

_protocolVersion ∷ !(HKD f ProtVer)

Protocol version

_minUTxOValue ∷ !(HKD f Coin)

Minimum UTxO value

_minPoolCost ∷ !(HKD f Coin)

Minimum Stake Pool Cost

data OCert era

Constructors

OCert

Fields

ocertVkHot ∷ !(VerKeyKES era)

The operational hot key

ocertN ∷ !Word64

counter

ocertKESPeriod ∷ !KESPeriod

Start of key evolving signature period

ocertSigma ∷ !(SignedDSIGN era (OCertSignable era))

Signature of block operational certificate content

data OCertEnv era

Constructors

OCertEnv

Fields

ocertEnvStPools ∷ Set (KeyHash StakePool era)

ocertEnvGenDelegs ∷ Set (KeyHash GenesisDelegate era)

data Ptr

Pointer to a slot, transaction index and index in certificate list.

Constructors

Ptr !SlotNo !Ix !Ix

data StakeReference era

Constructors

StakeRefBase !(StakeCredential era)

StakeRefPtr !Ptr

StakeRefNull

data Credential (kr ∷ KeyRole) era

Script hash or key hash for a payment or a staking object.

Constructors

ScriptHashObj !(ScriptHash era)

KeyHashObj !(KeyHash kr era)

data RewardAcnt era

An account based address for rewards

Constructors

RewardAcnt

Fields

getRwdNetwork ∷ !Network

getRwdCred ∷ !(Credential Staking era)

data Addr era

An address for UTxO.

Constructors

Addr !Network !(PaymentCredential era) !(StakeReference era)

AddrBootstrap !(BootstrapAddress era)

newtype StakeCreds era

Constructors

StakeCreds

Fields

unStakeCreds ∷ Map (Credential Staking era) SlotNo

data WitVKey era kr where

Proof/Witness that a transaction is authorized by the given key holder.

Bundled Patterns

pattern WitVKey ∷ (Typeable kr, Era era) ⇒ VKey kr era → SignedDSIGN era (Hash era (TxBody era)) → WitVKey era kr

data TxBody era where

A raw transaction

Bundled Patterns

pattern TxBody ∷ Era era ⇒ Set (TxIn era) → StrictSeq (TxOut era) → StrictSeq (DCert era) → Wdrl era → Coin → SlotNo → StrictMaybe (Update era) → StrictMaybe (MetaDataHash era) → TxBody era

data DCert era

A heavyweight certificate.

Constructors

DCertDeleg !(DelegCert era)

DCertPool !(PoolCert era)

DCertGenesis !(GenesisDelegCert era)

DCertMir !(MIRCert era)

data MIRCert era

Move instantaneous rewards certificate

Constructors

MIRCert

Fields

mirPot ∷ MIRPot

mirRewards ∷ Map (Credential Staking era) Coin

data MIRPot

Constructors

ReservesMIR

TreasuryMIR

data GenesisDelegCert era

Genesis key delegation certificate

Constructors

GenesisDelegCert !(KeyHash Genesis era) !(KeyHash GenesisDelegate era) !(Hash era (VerKeyVRF era))

data PoolCert era

Constructors

RegPool !(PoolParams era)
A stake pool registration certificate.

RetirePool !(KeyHash StakePool era) !EpochNo
A stake pool retirement certificate.

data DelegCert era

Constructors

RegKey !(StakeCredential era)
A stake key registration certificate.

DeRegKey !(StakeCredential era)
A stake key deregistration certificate.

Delegate !(Delegation era)
A stake delegation certificate.

data TxOut era

The output of a UTxO.

Constructors

TxOutCompact !ShortByteString !Word64

Bundled Patterns

pattern TxOut ∷ Era era ⇒ Addr era → Coin → TxOut era

data TxIn era

The input of a UTxO.

Constructors

TxInCompact !(TxId era) !Word64

Bundled Patterns

pattern TxIn ∷ Era era ⇒ TxId era → Natural → TxIn era

newtype TxId era

A unique ID of a transaction, which is computable from the transaction.

Constructors

TxId

Fields

_unTxId ∷ Hash era (TxBody era)

newtype Wdrl era

Constructors

Wdrl

Fields

unWdrl ∷ Map (RewardAcnt era) Coin

data PoolParams era

A stake pool.

Constructors

PoolParams

Fields

_poolPubKey ∷ !(KeyHash StakePool era)

_poolVrf ∷ !(Hash era (VerKeyVRF era))

_poolPledge ∷ !Coin

_poolCost ∷ !Coin

_poolMargin ∷ !UnitInterval

_poolRAcnt ∷ !(RewardAcnt era)

_poolOwners ∷ !(Set (KeyHash Staking era))

_poolRelays ∷ !(StrictSeq StakePoolRelay)

_poolMD ∷ !(StrictMaybe PoolMetaData)

data StakePoolRelay

Constructors

SingleHostAddr !(StrictMaybe Port) !(StrictMaybe IPv4) !(StrictMaybe IPv6)
One or both of IPv4 & IPv6

SingleHostName !(StrictMaybe Port) !DnsName
An A or AAAA DNS record

MultiHostName !DnsName
A SRV DNS record

data PoolMetaData

Constructors

PoolMetaData

Fields

_poolMDUrl ∷ !Url

_poolMDHash ∷ !ByteString

data Delegation era

The delegation of one stake key to another.

Constructors

Delegation

Fields

_delegator ∷ !(StakeCredential era)

_delegatee ∷ !(KeyHash StakePool era)

newtype PoolDistr era

Constructors

PoolDistr

Fields

unPoolDistr ∷ Map (KeyHash StakePool era) (IndividualPoolStake era)

data BootstrapWitness era where

Bundled Patterns

pattern BootstrapWitness ∷ Era era ⇒ VKey Witness era → SignedDSIGN era (Hash era (TxBody era)) → ChainCode → ByteString → BootstrapWitness era

data Tx era where

A fully formed transaction.

Bundled Patterns

pattern Tx ∷ Era era ⇒ TxBody era → WitnessSet era → StrictMaybe MetaData → Tx era

type WitnessSet = WitnessSetHKD Identity

data SnapShots era

Snapshots of the stake distribution.

Constructors

SnapShots !(SnapShot era) !(SnapShot era) !(SnapShot era) !Coin

data SnapShot era

Snapshot of the stake distribution.

Constructors

SnapShot !(Stake era) !(Map (Credential Staking era) (KeyHash StakePool era)) !(Map (KeyHash StakePool era) (PoolParams era))

newtype Stake era

Type of stake as map from hash key to coins associated.

Constructors

Stake (Map (Credential Staking era) Coin)

data NonMyopic era

newtype WitHashes era

Constructors

WitHashes

Fields

unWitHashes ∷ Set (KeyHash Witness era)

data LedgerState era

The state associated with a Ledger.

Constructors

LedgerState

Fields

_utxoState ∷ !(UTxOState era)

The current unspent transaction outputs.

_delegationState ∷ !(DPState era)

The current delegation state

data NewEpochState era

New Epoch state and environment

Constructors

NewEpochState

Fields

nesEL ∷ !EpochNo

Last epoch

nesBprev ∷ !(BlocksMade era)

Blocks made before current epoch

nesBcur ∷ !(BlocksMade era)

Blocks made in current epoch

nesEs ∷ !(EpochState era)

Epoch state before current

nesRu ∷ !(StrictMaybe (RewardUpdate era))

Possible reward update

nesPd ∷ !(PoolDistr era)

Stake distribution within the stake pool

data UTxOState era

Constructors

UTxOState

Fields

_utxo ∷ !(UTxO era)

_deposited ∷ !Coin

_fees ∷ !Coin

_ppups ∷ !(PPUPState era)

data EpochState era

Constructors

EpochState

Fields

esAccountState ∷ !AccountState

esSnapshots ∷ !(SnapShots era)

esLState ∷ !(LedgerState era)

esPrevPp ∷ !(PParams era)

esPp ∷ !(PParams era)

esNonMyopic ∷ !(NonMyopic era)

data AccountState

Constructors

AccountState

Fields

_treasury ∷ !Coin

_reserves ∷ !Coin

data RewardUpdate era

Constructors

RewardUpdate

Fields

deltaT ∷ !Coin

deltaR ∷ !Coin

rs ∷ !(Map (Credential Staking era) Coin)

deltaF ∷ !Coin

nonMyopic ∷ !(NonMyopic era)

data DPState era

The state associated with the current stake delegation.

Constructors

DPState

Fields

_dstate ∷ !(DState era)

_pstate ∷ !(PState era)

data PState era

Current state of staking pools and their certificate counters.

Constructors

PState

Fields

_pParams ∷ !(Map (KeyHash StakePool era) (PoolParams era))

The pool parameters.

_fPParams ∷ !(Map (KeyHash StakePool era) (PoolParams era))

The future pool parameters.

_retiring ∷ !(Map (KeyHash StakePool era) EpochNo)

A map of retiring stake pools to the epoch when they retire.

data DState era

State of staking pool delegations and rewards

Constructors

DState

Fields

_rewards ∷ !(RewardAccounts era)

The active reward accounts.

_delegations ∷ !(Map (Credential Staking era) (KeyHash StakePool era))

The current delegations.

_ptrs ∷ !(Bimap Ptr (Credential Staking era))

The pointed to hash keys.

_fGenDelegs ∷ !(Map (FutureGenDeleg era) (GenDelegPair era))

future genesis key delegations

_genDelegs ∷ !(GenDelegs era)

Genesis key delegations

_irwd ∷ !(InstantaneousRewards era)

Instantaneous Rewards

type KeyPairs era = [(KeyPair Payment era, KeyPair Staking era)]

Representation of a list of pairs of key pairs, e.g., pay and stake keys

data PPUPEnv era

Constructors

PPUPEnv SlotNo (PParams era) (GenDelegs era)

data PPUP era

data UtxoEnv era

Constructors

UtxoEnv SlotNo (PParams era) (Map (KeyHash StakePool era) (PoolParams era)) (GenDelegs era)

data UTXO era

data UTXOW era

data POOLREAP era

data PoolEnv era

Constructors

PoolEnv SlotNo (PParams era)

data POOL (era ∷ Type)

data NEWEPOCH era

data TICK era

data DelegEnv

Constructors

DelegEnv

Fields

slotNo ∷ SlotNo

ptr_ ∷ Ptr

acnt_ ∷ AccountState

data DELEG era

data DelplEnv era

Constructors

DelplEnv

Fields

delplSlotNo ∷ SlotNo

delPlPtr ∷ Ptr

delPlPp ∷ PParams era

delPlAcnt ∷ AccountState

data DELPL era

data DelegsEnv era

Constructors

DelegsEnv

Fields

delegsSlotNo ∷ SlotNo

delegsIx ∷ Ix

delegspp ∷ PParams era

delegsTx ∷ Tx era

delegsAccount ∷ AccountState

data DELEGS era

data LedgerEnv era

Constructors

LedgerEnv

Fields

ledgerSlotNo ∷ SlotNo

ledgerIx ∷ Ix

ledgerPp ∷ PParams era

ledgerAccount ∷ AccountState

data LEDGER era

data LedgersEnv era

Constructors

LedgersEnv

Fields

ledgersSlotNo ∷ SlotNo

ledgersPp ∷ PParams era

ledgersAccount ∷ AccountState

data LEDGERS era

newtype LaxBlock era

Constructors

LaxBlock (Block era)

data Block era where

Bundled Patterns

pattern Block ∷ Era era ⇒ BHeader era → TxSeq era → Block era

data BHBody era

Constructors

BHBody

Fields

bheaderBlockNo ∷ !BlockNo

block number

bheaderSlotNo ∷ !SlotNo

block slot

bheaderPrev ∷ !(PrevHash era)

Hash of the previous block header

bheaderVk ∷ !(VKey BlockIssuer era)

verification key of block issuer

bheaderVrfVk ∷ !(VerKeyVRF era)

VRF verification key for block issuer

bheaderEta ∷ !(CertifiedVRF era Nonce)

block nonce

bheaderL ∷ !(CertifiedVRF era Natural)

leader election value

bsize ∷ !Natural

Size of the block body

bhash ∷ !(HashBBody era)

Hash of block body

bheaderOCert ∷ !(OCert era)

operational certificate

bprotver ∷ !ProtVer

protocol version

data PrevHash era

The previous hash of a block

Constructors

GenesisHash

BlockHash !(HashHeader era)

data BHeader era where

Bundled Patterns

pattern BHeader ∷ Era era ⇒ BHBody era → SignedKES era (BHBody era) → BHeader era

newtype HashHeader era

The hash of a Block Header

Constructors

HashHeader

Fields

unHashHeader ∷ Hash era (BHeader era)

data ChainState era

Constructors

ChainState

Fields

chainNes ∷ NewEpochState era

chainOCertIssue ∷ Map (KeyHash BlockIssuer era) Word64

chainEpochNonce ∷ Nonce

chainEvolvingNonce ∷ Nonce

chainCandidateNonce ∷ Nonce

chainPrevEpochNonce ∷ Nonce

chainLastAppliedBlock ∷ WithOrigin (LastAppliedBlock era)

data CHAIN era