Shelley.Spec.Ledger.API.Validation

Description

Interface to the block validation and chain extension logic in the Shelley API.

Documentation

type ShelleyState = NewEpochState

Type alias for the state updated by TICK and BBODY rules

newtype TickTransitionError crypto

Constructors

TickTransitionError [PredicateFailure (TICK crypto)]

newtype BlockTransitionError crypto

Constructors

BlockTransitionError [PredicateFailure (BBODY crypto)]

chainChecks ∷ ∀ crypto m. (Crypto crypto, MonadError (PredicateFailure (CHAIN crypto)) m) ⇒ Globals → PParams → BHeader crypto → m ()

applyTickTransition ∷ ∀ crypto. Crypto crypto ⇒ Globals → ShelleyState crypto → SlotNo → ShelleyState crypto

Apply the header level ledger transition.

This handles checks and updates that happen on a slot tick, as well as a few header level checks, such as size constraints.

applyBlockTransition ∷ ∀ crypto m. (Crypto crypto, MonadError (BlockTransitionError crypto) m, DSignable crypto (Hash crypto (TxBody crypto))) ⇒ Globals → ShelleyState crypto → Block crypto → m (ShelleyState crypto)

Apply the block level ledger transition.