Shelley.Spec.Ledger.API.Wallet

Documentation

getNonMyopicMemberRewards ∷ Globals → ShelleyState crypto → Set (Credential Staking crypto) → Map (Credential Staking crypto) (Map (KeyHash StakePool crypto) Coin)

Calculate the Non-Myopic Pool Member Rewards for a set of credentials. For each given credential, this function returns a map from each stake pool (identified by the key hash of the pool operator) to the non-myopic pool member reward for that stake pool.

getUTxO ∷ ShelleyState crypto → UTxO crypto

Get the full UTxO.

getFilteredUTxO ∷ ShelleyState crypto → Set (Addr crypto) → UTxO crypto

Get the UTxO filtered by address.