Shelley.Spec.Ledger.Address

Documentation

mkVKeyRwdAcnt ∷ Crypto crypto ⇒ KeyPair Regular crypto → RewardAcnt crypto

mkRwdAcnt ∷ Credential crypto → RewardAcnt crypto

scriptsToAddr ∷ Crypto crypto ⇒ (MultiSig crypto, MultiSig crypto) → Addr crypto

Create a base address from a pair of multi-sig scripts (pay and stake)

scriptToCred ∷ Crypto crypto ⇒ MultiSig crypto → Credential crypto

Convert a given multi-sig script to a credential by hashing it and wrapping into the Credential data type.

toAddr ∷ Crypto crypto ⇒ (KeyPair Regular crypto, KeyPair Regular crypto) → Addr crypto

toCred ∷ Crypto crypto ⇒ KeyPair Regular crypto → Credential crypto

serialiseAddr ∷ Crypto crypto ⇒ Addr crypto → ByteString

Serialise an address to the external format.

Note: this is not the final format, this function will be updated later to use the final format.

See https://github.com/input-output-hk/cardano-ledger-specs/issues/1367

deserialiseAddr ∷ Crypto crypto ⇒ ByteString → Maybe (Addr crypto)

Deserialise an address from the external format. This will fail if the input data is not in the right format (or if there is trailing data).

Note: this is not the final format, this function will be updated later to use the final format.

See https://github.com/input-output-hk/cardano-ledger-specs/issues/1367