Shelley.Spec.Ledger.Address

Documentation

mkVKeyRwdAcnt ∷ Crypto crypto ⇒ KeyPair Regular crypto → RewardAcnt crypto

mkRwdAcnt ∷ Credential crypto → RewardAcnt crypto

scriptsToAddr ∷ Crypto crypto ⇒ (MultiSig crypto, MultiSig crypto) → Addr crypto

Create a base address from a pair of multi-sig scripts (pay and stake)

scriptToCred ∷ Crypto crypto ⇒ MultiSig crypto → Credential crypto

Convert a given multi-sig script to a credential by hashing it and wrapping into the Credential data type.

toAddr ∷ Crypto crypto ⇒ (KeyPair Regular crypto, KeyPair Regular crypto) → Addr crypto

toCred ∷ Crypto crypto ⇒ KeyPair Regular crypto → Credential crypto