Shelley.Spec.Ledger.BaseTypes

Documentation

type FixedPoint = Digits34

(==>) ∷ Bool → Bool → Bool infix 1

(⭒) ∷ Nonce → Nonce → Nonce

Evolve the nonce

data Nonce

Evolving nonce type.

Constructors

Nonce !(Hash SHA256 Nonce)

NeutralNonce
Identity element

newtype Seed

Seed to the verifiable random function.

We do not expose the constructor to Seed. Instead, a Seed should be created using mkSeed for a VRF calculation.

Constructors

Seed (Hash SHA256 Seed)

newtype UnitInterval

Type to represent a value in the unit interval [0; 1]

Constructors

UnsafeUnitInterval Rational

fpEpsilon ∷ FixedPoint

fpPrecision ∷ FixedPoint

interval0 ∷ UnitInterval

interval1 ∷ UnitInterval

intervalValue ∷ UnitInterval → Rational

Get rational value of UnitInterval type

invalidKey ∷ Word → Decoder s a

mkNonce ∷ Natural → Nonce

Make a nonce from a natural number

mkUnitInterval ∷ Rational → Maybe UnitInterval

Return a UnitInterval type if r is in [0; 1].

truncateUnitInterval ∷ Rational → UnitInterval

Convert a rational to a UnitInterval by ignoring its integer part.

data StrictMaybe a

Strict Maybe.

TODO move to cardano-prelude

Constructors

SNothing

SJust !a

strictMaybeToMaybe ∷ StrictMaybe a → Maybe a

maybeToStrictMaybe ∷ Maybe a → StrictMaybe a

fromSMaybe ∷ a → StrictMaybe a → a

data Url

urlToText ∷ Url → Text

textToUrl ∷ Text → Maybe Url

data DnsName

dnsToText ∷ DnsName → Text

textToDns ∷ Text → Maybe DnsName

data Port

portToWord16 ∷ Port → Word16

data ActiveSlotCoeff

mkActiveSlotCoeff ∷ UnitInterval → ActiveSlotCoeff

activeSlotVal ∷ ActiveSlotCoeff → UnitInterval

activeSlotLog ∷ ActiveSlotCoeff → FixedPoint

STS Base

data Globals

Constructors

Globals

Fields

epochInfo ∷ !(EpochInfo Identity)

slotsPerKESPeriod ∷ !Word64

slotsPrior ∷ !Word64

Number of slots before the end of the epoch at which we stop updating the candidate nonce for the next epoch.

This value is also used in a number of other places; for example, protocol updates must be submitted at least this many slots before an epoch boundary.

startRewards ∷ !Word64

Number of slots after the beginning of an epoch when we may begin to distribute rewards.

securityParameter ∷ !Word64

Maximum number of blocks we are allowed to roll back

maxKESEvo ∷ !Word64

Maximum number of KES iterations

quorum ∷ !Word64

Quorum for update system votes and MIR certificates

maxMajorPV ∷ !Natural

All blocks invalid after this protocol version

maxLovelaceSupply ∷ !Word64

Maximum number of lovelace in the system

activeSlotCoeff ∷ !ActiveSlotCoeff

Active Slot Coefficient, named f in "Ouroboros Praos: An adaptively-secure, semi-synchronous proof-of-stake protocol"

type ShelleyBase = ReaderT Globals Identity