Shelley.Spec.Ledger.BlockChain

Documentation

newtype HashHeader crypto

The hash of a Block Header

Constructors

HashHeader

Fields

unHashHeader ∷ Hash crypto (BHeader crypto)

data PrevHash crypto

The previous hash of a block

Constructors

GenesisHash

BlockHash !(HashHeader crypto)

data LastAppliedBlock crypto

Constructors

LastAppliedBlock

Fields

labBlockNo ∷ !BlockNo

labSlotNo ∷ !SlotNo

labHash ∷ !(HashHeader crypto)

lastAppliedHash ∷ WithOrigin (LastAppliedBlock crypto) → PrevHash crypto

data BHBody crypto

Constructors

BHBody

Fields

bheaderBlockNo ∷ !BlockNo

block number

bheaderSlotNo ∷ !SlotNo

block slot

bheaderPrev ∷ !(PrevHash crypto)

Hash of the previous block header

bheaderVk ∷ !(VKey BlockIssuer crypto)

verification key of block issuer

bheaderVrfVk ∷ !(VerKeyVRF crypto)

VRF verification key for block issuer

bheaderEta ∷ !(CertifiedVRF crypto Nonce)

block nonce

bheaderL ∷ !(CertifiedVRF crypto UnitInterval)

leader election value

bsize ∷ !Natural

Size of the block body

bhash ∷ !(HashBBody crypto)

Hash of block body

bheaderOCert ∷ !(OCert crypto)

operational certificate

bprotver ∷ !ProtVer

protocol version

data BHeader crypto where

Bundled Patterns

pattern BHeader ∷ Crypto crypto ⇒ BHBody crypto → SignedKES crypto (BHBody crypto) → BHeader crypto

data Block crypto where

Bundled Patterns

pattern Block ∷ Crypto crypto ⇒ BHeader crypto → TxSeq crypto → Block crypto

newtype LaxBlock crypto

Constructors

LaxBlock (Block crypto)

data TxSeq crypto where

Bundled Patterns

pattern TxSeq ∷ Crypto crypto ⇒ StrictSeq (Tx crypto) → TxSeq crypto

data HashBBody crypto

Hash of block body

bhHash ∷ Crypto crypto ⇒ BHeader crypto → HashHeader crypto

Hash a given block header

bbHash ∷ ∀ crypto. Crypto crypto ⇒ TxSeq crypto → HashBBody crypto

Hash a given block body

hashHeaderToNonce ∷ HashHeader crypto → Nonce

HashHeader to Nonce

prevHashToNonce ∷ PrevHash crypto → Nonce

bHeaderSize ∷ ∀ crypto. Crypto crypto ⇒ BHeader crypto → Int

bBodySize ∷ ∀ crypto. Crypto crypto ⇒ TxSeq crypto → Int

slotToNonce ∷ SlotNo → Nonce

hBbsize ∷ BHBody crypto → Natural

bheader ∷ Crypto crypto ⇒ Block crypto → BHeader crypto

bhbody ∷ Crypto crypto ⇒ BHeader crypto → BHBody crypto

bbody ∷ Crypto crypto ⇒ Block crypto → TxSeq crypto

hsig ∷ Crypto crypto ⇒ BHeader crypto → SignedKES crypto (BHBody crypto)

seedEta ∷ Nonce

seedL ∷ Nonce

incrBlocks ∷ Bool → KeyHash StakePool crypto → BlocksMade crypto → BlocksMade crypto

mkSeed

Arguments

∷ Nonce
Universal constant

→ SlotNo

→ Nonce
Epoch nonce

→ Seed

Construct a seed to use in the VRF computation.

checkVRFValue ∷ Natural → Rational → ActiveSlotCoeff → Bool

Check that the certified input natural is valid for being slot leader. This means we check that

fromNat (certNat) < 1 - (1 - f)^σ

where fromNat creates an appropriate value in [0;1] from the certified natural. The calculation is done using the following optimization:

let p = fromNat (certNat) and c = ln(1 - f)

then p < 1 - (1 - f)^σ = 1 / (1 - p) > exp(-σ * c)

this can be efficiently be computed by taylorExpCmp which returns ABOVE in case the reference value `1 / (1 - p)` is above the exponential function at `-σ * c`, BELOW if it is below or MaxReached if it couldn't conclusively compute this within the given iteration bounds.