Shelley.Spec.Ledger.Credential

Documentation

data Credential (kr ∷ KeyRole) crypto

Script hash or key hash for a payment or a staking object.

Constructors

ScriptHashObj !(ScriptHash crypto)

KeyHashObj !(KeyHash kr crypto)

newtype GenesisCredential crypto

Constructors

GenesisCredential (KeyHash Genesis crypto)

type Ix = Natural

type PaymentCredential crypto = Credential Payment crypto

data Ptr

Pointer to a slot, transaction index and index in certificate list.

Constructors

Ptr !SlotNo !Ix !Ix

newtype RewardAcnt crypto

An account based address for rewards

Constructors

RewardAcnt

Fields

getRwdCred ∷ Credential Staking crypto

type StakeCredential crypto = Credential Staking crypto

data StakeReference crypto

Constructors

StakeRefBase !(StakeCredential crypto)

StakeRefPtr !Ptr

StakeRefNull