Shelley.Spec.Ledger.Delegation.Certificates

Documentation

data DCert crypto

A heavyweight certificate.

Constructors

DCertDeleg (DelegCert crypto)

DCertPool (PoolCert crypto)

DCertGenesis (GenesisDelegate crypto)

DCertMir (MIRCert crypto)

data DelegCert crypto

Constructors

RegKey (Credential crypto)
A stake key registration certificate.

DeRegKey (Credential crypto)
A stake key deregistration certificate.

Delegate (Delegation crypto)
A stake delegation certificate.

data PoolCert crypto

Constructors

RegPool (PoolParams crypto)
A stake pool registration certificate.

RetirePool (KeyHash crypto) EpochNo
A stake pool retirement certificate.

newtype GenesisDelegate crypto

Genesis key delegation certificate

Constructors

GenesisDelegate (GenKeyHash crypto, KeyHash crypto)

newtype MIRCert crypto

Move instantaneous rewards certificate

Constructors

MIRCert (Map (Credential crypto) Coin)

newtype StakeCreds crypto

Constructors

StakeCreds (Map (Credential crypto) SlotNo)

newtype StakePools crypto

Constructors

StakePools

Fields

unStakePools ∷ Map (KeyHash crypto) SlotNo

newtype PoolDistr crypto

Constructors

PoolDistr

Fields

unPoolDistr ∷ Map (KeyHash crypto) (Rational, Hash (HASH crypto) (VerKeyVRF (VRF crypto)))

delegCWitness ∷ DelegCert crypto → Credential crypto

Determine the certificate author

poolCWitness ∷ PoolCert crypto → Credential crypto

genesisCWitness ∷ GenesisDelegate crypto → GenKeyHash crypto

dvalue ∷ DCert crypto → PParams → Coin

Retrieve the deposit amount for a certificate

refund ∷ Coin → UnitInterval → Rational → Duration → Coin

Compute a refund on a deposit

decayKey ∷ PParams → (Coin, UnitInterval, Rational)

decayPool ∷ PParams → (Coin, UnitInterval, Rational)

isRegKey ∷ DCert crypto → Bool

Check for RegKey constructor

isDeRegKey ∷ DCert crypto → Bool

Check for DeRegKey constructor

isDelegation ∷ DCert crypto → Bool

Check for Delegation constructor

isGenesisDelegation ∷ DCert crypto → Bool

Check for GenesisDelegate constructor

isRegPool ∷ DCert crypto → Bool

Check for RegPool constructor

isRetirePool ∷ DCert crypto → Bool

Check for RetirePool constructor

isInstantaneousRewards ∷ DCert crypto → Bool

requiresVKeyWitness ∷ DCert crypto → Bool

Returns True for delegation certificates that require at least one witness, and False otherwise. It is mainly used to ensure that calling a variant of cwitness is safe.