Shelley.Spec.Ledger.EpochBoundary

Description

This modules implements the necessary functions for the changes that can happen at epoch boundaries.

Documentation

newtype Stake crypto

Type of stake as map from hash key to coins associated.

Constructors

Stake

Fields

unStake ∷ Map (Credential Staking crypto) Coin

newtype BlocksMade crypto

Blocks made

Constructors

BlocksMade (Map (KeyHash StakePool crypto) Natural)

data SnapShot crypto

Snapshot of the stake distribution.

Constructors

SnapShot

Fields

_stake ∷ !(Stake crypto)

_delegations ∷ !(Map (Credential Staking crypto) (KeyHash StakePool crypto))

_poolParams ∷ !(Map (KeyHash StakePool crypto) (PoolParams crypto))

data SnapShots crypto

Snapshots of the stake distribution.

Constructors

SnapShots

Fields

_pstakeMark ∷ !(SnapShot crypto)

_pstakeSet ∷ !(SnapShot crypto)

_pstakeGo ∷ !(SnapShot crypto)

_feeSS ∷ !Coin

emptySnapShot ∷ SnapShot crypto

emptySnapShots ∷ SnapShots crypto

rewardStake ∷ ∀ crypto. Map (RewardAcnt crypto) Coin → [(Credential Staking crypto, Coin)]

aggregateOuts ∷ UTxO crypto → Map (Addr crypto) Coin

baseStake ∷ Map (Addr crypto) Coin → [(Credential Staking crypto, Coin)]

Get Stake of base addresses in TxOut set.

ptrStake ∷ ∀ crypto. Map (Addr crypto) Coin → Map Ptr (Credential Staking crypto) → [(Credential Staking crypto, Coin)]

Calculate stake of pointer addresses in TxOut set.

poolStake ∷ KeyHash StakePool crypto → Map (Credential Staking crypto) (KeyHash StakePool crypto) → Stake crypto → Stake crypto

Get stake of one pool

obligation ∷ PParams → StakeCreds crypto → StakePools crypto → SlotNo → Coin

Calculate total possible refunds.

poolRefunds ∷ PParams → Map (KeyHash StakePool crypto) SlotNo → SlotNo → Map (KeyHash StakePool crypto) Coin

Calculate pool refunds

maxPool ∷ PParams → Coin → Rational → Rational → Coin

Calculate maximal pool reward

groupByPool ∷ Map (KeyHash Staking crypto) Coin → Map (KeyHash Staking crypto) (KeyHash StakePool crypto) → Map (KeyHash StakePool crypto) (Map (KeyHash Staking crypto) Coin)

Pool individual reward

(⊎) ∷ Stake crypto → Stake crypto → Stake crypto

Add two stake distributions