Shelley.Spec.Ledger.Keys

Documentation

data KeyDiscriminator

Discriminate between keys based on their usage in the system.

Constructors

Genesis

Regular

class Typeable h ⇒ HashAlgorithm h

Minimal complete definition

byteCount, digest

data Hash h a

hash ∷ (HashAlgorithm h, ToCBOR a) ⇒ a → Hash h a

class (Typeable v, Show (VerKeyDSIGN v), Eq (VerKeyDSIGN v), Show (SignKeyDSIGN v), Show (SigDSIGN v), Eq (SigDSIGN v), NoUnexpectedThunks (SigDSIGN v), NoUnexpectedThunks (SignKeyDSIGN v), NoUnexpectedThunks (VerKeyDSIGN v)) ⇒ DSIGNAlgorithm v

Minimal complete definition

abstractSizeVKey, abstractSizeSig, encodeVerKeyDSIGN, encodeSignKeyDSIGN, encodeSigDSIGN, decodeVerKeyDSIGN, decodeSignKeyDSIGN, decodeSigDSIGN, genKeyDSIGN, deriveVerKeyDSIGN, signDSIGN, verifyDSIGN

type family Signable v ∷ Type → Constraint

newtype SKey crypto

Constructors

SKey (SignKeyDSIGN (DSIGN crypto))

newtype DiscVKey (kd ∷ KeyDiscriminator) crypto

Discriminated verification key

Constructors

DiscVKey (VerKeyDSIGN (DSIGN crypto))

type VKey = DiscVKey Regular

pattern VKey ∷ VerKeyDSIGN (DSIGN crypto) → DiscVKey Regular crypto

type VKeyGenesis = DiscVKey Genesis

pattern VKeyGenesis ∷ VerKeyDSIGN (DSIGN crypto) → DiscVKey Genesis crypto

newtype DiscKeyHash (discriminator ∷ KeyDiscriminator) crypto

Discriminated hash of public Key

Constructors

DiscKeyHash (Hash (HASH crypto) (VerKeyDSIGN (DSIGN crypto)))

type KeyHash crypto = DiscKeyHash Regular crypto

pattern KeyHash ∷ Hash (HASH crypto) (VerKeyDSIGN (DSIGN crypto)) → DiscKeyHash Regular crypto

type GenKeyHash crypto = DiscKeyHash Genesis crypto

pattern GenKeyHash ∷ Hash (HASH crypto) (VerKeyDSIGN (DSIGN crypto)) → DiscKeyHash Genesis crypto

data AnyKeyHash crypto

Discriminated hash of public Key

pattern AnyKeyHash ∷ Hash (HASH crypto) (VerKeyDSIGN (DSIGN crypto)) → AnyKeyHash crypto

undiscriminateKeyHash ∷ DiscKeyHash kd crypto → AnyKeyHash crypto

data KeyPair (kd ∷ KeyDiscriminator) crypto

Constructors

KeyPair

Fields

vKey ∷ DiscVKey kd crypto

sKey ∷ SKey crypto

newtype Sig crypto a

Constructors

UnsafeSig (SignedDSIGN (DSIGN crypto) a)

hashKey ∷ Crypto crypto ⇒ DiscVKey kd crypto → DiscKeyHash kd crypto

Hash a given public key

hashAnyKey ∷ Crypto crypto ⇒ DiscVKey kd crypto → AnyKeyHash crypto

Hash a given public key and forget about what it's a hash of

sign ∷ (Crypto crypto, Signable (DSIGN crypto) a) ⇒ SKey crypto → a → Sig crypto a

Produce a digital signature

verify ∷ (Crypto crypto, Signable (DSIGN crypto) a) ⇒ DiscVKey kd crypto → a → Sig crypto a → Bool

Verify a digital signature

newtype GKeys crypto

Constructors

GKeys (Set (VKeyGenesis crypto))

newtype GenDelegs crypto

Constructors

GenDelegs (Map (GenKeyHash crypto) (KeyHash crypto))

class (Typeable v, Show (VerKeyKES v), Eq (VerKeyKES v), Show (SignKeyKES v), Show (SigKES v), Eq (SigKES v), NoUnexpectedThunks (SigKES v), NoUnexpectedThunks (SignKeyKES v), NoUnexpectedThunks (VerKeyKES v)) ⇒ KESAlgorithm v

Minimal complete definition

encodeVerKeyKES, encodeSignKeyKES, encodeSigKES, decodeVerKeyKES, decodeSignKeyKES, decodeSigKES, genKeyKES, deriveVerKeyKES, updateKES, signKES, verifyKES, currentPeriodKES

type KESignable crypto a = Signable (KES crypto) a

newtype SKeyES crypto

Constructors

SKeyES (SignKeyKES (KES crypto))

newtype VKeyES crypto

Constructors

VKeyES (VerKeyKES (KES crypto))

newtype KeyHashES crypto

The hash of KES verification Key

Constructors

KeyHashES (Hash (HASH crypto) (VerKeyKES (KES crypto)))

newtype KESig crypto a

Constructors

KESig (SignedKES (KES crypto) a)

hashKeyES ∷ Crypto crypto ⇒ VKeyES crypto → KeyHashES crypto

Hash a given public key

signKES ∷ (Crypto crypto, Signable (KES crypto) a) ⇒ SKeyES crypto → a → Natural → Maybe (KESig crypto a)

Produce a key evolving signature

verifyKES ∷ (Crypto crypto, Signable (KES crypto) a) ⇒ VKeyES crypto → a → KESig crypto a → Natural → Bool

Verify a key evolving signature

updateKESKey

Arguments

∷ Crypto crypto

⇒ SKeyES crypto

→ Natural
Target period

→ Maybe (SKeyES crypto)

currentPeriodKESKey ∷ Crypto crypto ⇒ SKeyES crypto → Natural

class (Typeable v, Show (VerKeyVRF v), Eq (VerKeyVRF v), Show (SignKeyVRF v), Show (CertVRF v), Eq (CertVRF v), FromCBOR (CertVRF v), ToCBOR (CertVRF v), NoUnexpectedThunks (CertVRF v), NoUnexpectedThunks (VerKeyVRF v), NoUnexpectedThunks (SignKeyVRF v)) ⇒ VRFAlgorithm v

Minimal complete definition

maxVRF, genKeyVRF, deriveVerKeyVRF, decodeVerKeyVRF, encodeVerKeyVRF, evalVRF, verifyVRF

Associated Types

data VerKeyVRF v ∷ Type

data SignKeyVRF v ∷ Type

class VRFValue a where

Our VRFAlgorithm provides a Natural, so we must exhibit an extractor to use the bits as some other type

Methods

fromNatural ∷ Natural → a

Extract a value from a natural number derived from the VRF computation.

data CertifiedVRF v a

Constructors

CertifiedVRF

Fields

certifiedNatural ∷ Natural

certifiedProof ∷ CertVRF v

hashKeyVRF ∷ Crypto crypto ⇒ VerKeyVRF (VRF crypto) → Hash (HASH crypto) (VerKeyVRF (VRF crypto))