Shelley.Spec.Ledger.LedgerState

Description

This module implements the operation rules for treating UTxO transactions (Tx) as state transformations on a ledger state (LedgerState), as specified in A Simplified Formal Specification of a UTxO Ledger.

Documentation

data LedgerState crypto

The state associated with a Ledger.

Constructors

LedgerState

Fields

_utxoState ∷ !(UTxOState crypto)

The current unspent transaction outputs.

_delegationState ∷ !(DPState crypto)

The current delegation state

type Ix = Natural

data DPState crypto

The state associated with the current stake delegation.

Constructors

DPState

Fields

_dstate ∷ !(DState crypto)

_pstate ∷ !(PState crypto)

data DState crypto

State of staking pool delegations and rewards

Constructors

DState

Fields

_stkCreds ∷ !(StakeCreds crypto)

The active stake keys.

_rewards ∷ !(RewardAccounts crypto)

The active reward accounts.

_delegations ∷ !(Map (Credential crypto) (KeyHash crypto))

The current delegations.

_ptrs ∷ !(Map Ptr (Credential crypto))

The pointed to hash keys.

_fGenDelegs ∷ !(Map (FutureGenDeleg crypto) (KeyHash crypto))

future genesis key delegations

_genDelegs ∷ !(GenDelegs crypto)

Genesis key delegations

_irwd ∷ !(Map (Credential crypto) Coin)

Instantaneous Rewards

data AccountState

Constructors

AccountState

Fields

_treasury ∷ !Coin

_reserves ∷ !Coin

data RewardUpdate crypto

Constructors

RewardUpdate

Fields

deltaT ∷ !Coin

deltaR ∷ !Coin

rs ∷ !(Map (RewardAcnt crypto) Coin)

deltaF ∷ !Coin

nonMyopic ∷ !(NonMyopic crypto)

type RewardAccounts crypto = Map (RewardAcnt crypto) Coin

emptyRewardUpdate ∷ RewardUpdate crypto

data FutureGenDeleg crypto

Constructors

FutureGenDeleg

Fields

fGenDelegSlot ∷ !SlotNo

fGenDelegGenKeyHash ∷ !(GenKeyHash crypto)

data EpochState crypto

Constructors

EpochState

Fields

esAccountState ∷ !AccountState

esSnapshots ∷ !(SnapShots crypto)

esLState ∷ !(LedgerState crypto)

esPrevPp ∷ !PParams

esPp ∷ !PParams

esNonMyopic ∷ !(NonMyopic crypto)

emptyEpochState ∷ EpochState crypto

getIR ∷ EpochState crypto → Map (Credential crypto) Coin

emptyLedgerState ∷ LedgerState crypto

emptyUTxOState ∷ UTxOState crypto

clearPpup ∷ UTxOState crypto → UTxOState crypto

Clear the protocol parameter updates

data PState crypto

Current state of staking pools and their certificate counters.

Constructors

PState

Fields

_stPools ∷ !(StakePools crypto)

The active stake pools.

_pParams ∷ !(Map (KeyHash crypto) (PoolParams crypto))

The pool parameters.

_retiring ∷ !(Map (KeyHash crypto) EpochNo)

A map of retiring stake pools to the epoch when they retire.

type KeyPairs crypto = [(KeyPair Regular crypto, KeyPair Regular crypto)]

Representation of a list of pairs of key pairs, e.g., pay and stake keys

data UTxOState crypto

Constructors

UTxOState

Fields

_utxo ∷ !(UTxO crypto)

_deposited ∷ !Coin

_fees ∷ !Coin

_ppups ∷ !(ProposedPPUpdates crypto)

data OBftSlot crypto

Constructors

NonActiveSlot

ActiveSlot !(GenKeyHash crypto)

emptyAccount ∷ AccountState

emptyPState ∷ PState crypto

emptyDState ∷ DState crypto

state transitions

emptyDelegation ∷ DPState crypto

applyTxBody ∷ Crypto crypto ⇒ LedgerState crypto → PParams → TxBody crypto → LedgerState crypto

Apply a transaction body as a state transition function on the ledger state.

Genesis State

genesisId ∷ Crypto crypto ⇒ TxId crypto

The transaction Id for UTxO included at the beginning of a new ledger.

genesisCoins ∷ Crypto crypto ⇒ [TxOut crypto] → UTxO crypto

Creates the UTxO for a new ledger with the specified transaction outputs.

genesisState ∷ Map (GenKeyHash crypto) (KeyHash crypto) → UTxO crypto → LedgerState crypto

Creates the ledger state for an empty ledger which contains the specified transaction outputs.

Validation

data LedgerValidation crypto

A ledger validation state consists of a ledger state t and the list of validation errors that occurred from a valid s to reach t.

Constructors

LedgerValidation [ValidationError] (LedgerState crypto)

minfee ∷ ∀ crypto. Crypto crypto ⇒ PParams → Tx crypto → Coin

Minimum fee calculation

txsize ∷ ∀ crypto. Crypto crypto ⇒ Tx crypto → Integer

Implementation of abstract transaction size

validTx ∷ (Crypto crypto, Signable (DSIGN crypto) (TxBody crypto)) ⇒ Tx crypto → GenDelegs crypto → SlotNo → PParams → LedgerState crypto → Validity

produced ∷ Crypto crypto ⇒ PParams → StakePools crypto → TxBody crypto → Coin

Compute the lovelace which are created by the transaction

consumed ∷ Crypto crypto ⇒ PParams → UTxO crypto → StakeCreds crypto → TxBody crypto → Coin

Compute the lovelace which are destroyed by the transaction

verifiedWits ∷ (Crypto crypto, Signable (DSIGN crypto) (TxBody crypto)) ⇒ Tx crypto → Validity

Given a ledger state, determine if the UTxO witnesses in a given transaction are correct.

witsVKeyNeeded ∷ Crypto crypto ⇒ UTxO crypto → Tx crypto → GenDelegs crypto → Set (AnyKeyHash crypto)

Collect the set of hashes of keys that needs to sign a given transaction. This set consists of the txin owners, certificate authors, and withdrawal reward accounts.

keyRefunds ∷ Crypto crypto ⇒ PParams → StakeCreds crypto → TxBody crypto → Coin

Compute the key deregistration refunds in a transaction

keyRefund ∷ Coin → UnitInterval → Rational → StakeCreds crypto → SlotNo → DCert crypto → Coin

Key refund for a deregistration certificate.

decayedTx ∷ Crypto crypto ⇒ PParams → StakeCreds crypto → TxBody crypto → ShelleyBase Coin

Decayed deposit portions

stakeDistr ∷ ∀ crypto. UTxO crypto → DState crypto → PState crypto → SnapShot crypto

Stake distribution

applyRUpd ∷ RewardUpdate crypto → EpochState crypto → EpochState crypto

Apply a reward update

createRUpd ∷ EpochNo → BlocksMade crypto → EpochState crypto → Coin → ShelleyBase (RewardUpdate crypto)

Create a reward update

data NewEpochState crypto

New Epoch state and environment

Constructors

NewEpochState

Fields

nesEL ∷ !EpochNo

Last epoch

nesBprev ∷ !(BlocksMade crypto)

Blocks made before current epoch

nesBcur ∷ !(BlocksMade crypto)

Blocks made in current epoch

nesEs ∷ !(EpochState crypto)

Epoch state before current

nesRu ∷ !(StrictMaybe (RewardUpdate crypto))

Possible reward update

nesPd ∷ !(PoolDistr crypto)

Stake distribution within the stake pool

nesOsched ∷ !(Map SlotNo (OBftSlot crypto))

Overlay schedule for PBFT vs Praos

data NewEpochEnv crypto

Constructors

NewEpochEnv

Fields

neeS ∷ SlotNo

neeGkeys ∷ Set (GenKeyHash crypto)

overlaySchedule ∷ EpochNo → Set (GenKeyHash crypto) → PParams → ShelleyBase (Map SlotNo (OBftSlot crypto))

Overlay schedule This is just a very simple round-robin, evenly spaced schedule.

getGKeys ∷ NewEpochState crypto → Set (GenKeyHash crypto)

updateNES ∷ NewEpochState crypto → BlocksMade crypto → LedgerState crypto → NewEpochState crypto

Update new epoch state