Shelley.Spec.Ledger.OverlaySchedule

Overlay schedule

data OverlaySchedule crypto

compactOverlaySchedule ∷ OverlaySchedule crypto → Map (OBftSlot crypto) (NonEmpty SlotNo)

Convert the overlay schedule to a representation that is more compact when serialised to a bytestring, but less efficient for lookups.

Each genesis key hash will only be stored once, instead of each time it is assigned to a slot.

decompactOverlaySchedule ∷ Map (OBftSlot crypto) (NonEmpty SlotNo) → OverlaySchedule crypto

Inverse of compactOverlaySchedule

emptyOverlaySchedule ∷ OverlaySchedule crypto

isOverlaySlot ∷ SlotNo → OverlaySchedule c → Bool

lookupInOverlaySchedule ∷ SlotNo → OverlaySchedule crypto → Maybe (OBftSlot crypto)

overlaySchedule ∷ EpochNo → Set (KeyHash Genesis crypto) → PParams → ShelleyBase (OverlaySchedule crypto)

Overlay schedule This is just a very simple round-robin, evenly spaced schedule.

overlayScheduleHelper

Arguments

∷ EpochSize

→ SlotNo
First slot of the epoch

→ Set (KeyHash Genesis crypto)

→ UnitInterval
Decentralization parameter d

→ ActiveSlotCoeff

→ OverlaySchedule crypto

overlayScheduleIsEmpty ∷ OverlaySchedule crypto → Bool

overlayScheduleToMap ∷ OverlaySchedule crypto → Map SlotNo (OBftSlot crypto)

OBftSlot

data OBftSlot crypto

Constructors

NonActiveSlot

ActiveSlot !(KeyHash Genesis crypto)