Shelley.Spec.Ledger.OverlaySchedule

Overlay schedule

isOverlaySlot ∷ SlotNo → UnitInterval → SlotNo → Bool

classifyOverlaySlot ∷ SlotNo → [KeyHash Genesis era] → UnitInterval → ActiveSlotCoeff → SlotNo → OBftSlot era

lookupInOverlaySchedule ∷ SlotNo → [KeyHash Genesis era] → UnitInterval → ActiveSlotCoeff → SlotNo → Maybe (OBftSlot era)

OBftSlot

data OBftSlot era

Constructors

NonActiveSlot

ActiveSlot !(KeyHash Genesis era)

Testing

overlaySlots ∷ SlotNo → UnitInterval → EpochSize → [SlotNo]

Return the list of overlaySlots for a given epoch. Note that this linear in the size of the epoch, and should probably only be used for testing. If something more performant is needed, we could probably use [start + floor(xd) | x <- [0 .. (spe -1)], floor(xd) < spe] but we would need to make sure that this is equivalent.