Shelley.Spec.Ledger.STS.Prtcl

Documentation

data PRTCL crypto

data family PredicateFailure a ∷ Type

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

type family State a ∷ Type

Type of the state which the system transitions between.

data PrtclEnv crypto

Constructors

PrtclEnv PParams (Map SlotNo (Maybe (GenKeyHash crypto))) (PoolDistr crypto) (GenDelegs crypto) SlotNo Bool
New epoch marker

data PrtclState crypto

Constructors

PrtclState

Fields

(Map (KeyHash crypto) Natural)

(WithOrigin (LastAppliedBlock crypto))

Nonce

Current epoch nonce

Nonce

Evolving nonce

Nonce

Candidate nonce

Nonce

Prev epoch hash nonce