Shelley.Spec.Ledger.Scripts

Documentation

data MultiSig crypto where

Bundled Patterns

pattern RequireAllOf ∷ Crypto crypto ⇒ [MultiSig crypto] → MultiSig crypto

pattern RequireAnyOf ∷ Crypto crypto ⇒ [MultiSig crypto] → MultiSig crypto

pattern RequireSignature ∷ Crypto crypto ⇒ KeyHash Witness crypto → MultiSig crypto

pattern RequireMOf ∷ Crypto crypto ⇒ Int → [MultiSig crypto] → MultiSig crypto

data Script crypto

Constructors

MultiSigScript (MultiSig crypto)

newtype ScriptHash crypto

Constructors

ScriptHash (Hash crypto (Script crypto))

countMSigNodes ∷ Crypto crypto ⇒ MultiSig crypto → Int

Count nodes and leaves of multi signature script

getKeyCombination ∷ Crypto crypto ⇒ MultiSig crypto → [KeyHash Witness crypto]

Get one possible combination of keys for multi signature script

getKeyCombinations ∷ Crypto crypto ⇒ MultiSig crypto → [[KeyHash Witness crypto]]

Get all valid combinations of keys for given multi signature. This is mainly useful for testing.

hashAnyScript ∷ Crypto crypto ⇒ Script crypto → ScriptHash crypto

hashMultiSigScript ∷ Crypto crypto ⇒ MultiSig crypto → ScriptHash crypto

Hashes native multi-signature script, appending the nativeMultiSigTag in front and then calling the script CBOR function.