Shelley.Spec.Ledger.Tx

Documentation

data Tx era where

A fully formed transaction.

Bundled Patterns

pattern Tx ∷ Era era ⇒ TxBody era → WitnessSet era → StrictMaybe MetaData → Tx era

data TxBody era where

A raw transaction

Bundled Patterns

pattern TxBody ∷ Era era ⇒ Set (TxIn era) → StrictSeq (TxOut era) → StrictSeq (DCert era) → Wdrl era → Coin → SlotNo → StrictMaybe (Update era) → StrictMaybe (MetaDataHash era) → TxBody era

data TxOut era

The output of a UTxO.

Constructors

TxOutCompact !ShortByteString !Word64

Bundled Patterns

pattern TxOut ∷ Era era ⇒ Addr era → Coin → TxOut era

data TxIn era

The input of a UTxO.

Constructors

TxInCompact !(TxId era) !Word64

Bundled Patterns

pattern TxIn ∷ Era era ⇒ TxId era → Natural → TxIn era

newtype TxId era

A unique ID of a transaction, which is computable from the transaction.

Constructors

TxId

Fields

_unTxId ∷ Hash era (TxBody era)

decodeWits ∷ ∀ era s. Era era ⇒ Decoder s (Annotator (WitnessSet era))

segwitTx ∷ Era era ⇒ Annotator (TxBody era) → Annotator (WitnessSet era) → Maybe (Annotator MetaData) → Annotator (Tx era)

type WitnessSet = WitnessSetHKD Identity

data WitnessSetHKD f era where

Bundled Patterns

pattern WitnessSet ∷ Era era ⇒ Set (WitVKey era Witness) → Map (ScriptHash era) (MultiSig era) → Set (BootstrapWitness era) → WitnessSet era

data WitVKey era kr where

Proof/Witness that a transaction is authorized by the given key holder.

Bundled Patterns

pattern WitVKey ∷ (Typeable kr, Era era) ⇒ VKey kr era → SignedDSIGN era (Hash era (TxBody era)) → WitVKey era kr

class (Era era, ToCBOR a) ⇒ MultiSignatureScript a era

Typeclass for multis-signature script data types. Allows for script validation and hashing.

Minimal complete definition

validateScript, hashScript

validateScript ∷ MultiSignatureScript a era ⇒ a → Tx era → Bool

hashScript ∷ MultiSignatureScript a era ⇒ a → ScriptHash era

txwitsScript ∷ Era era ⇒ Tx era → Map (ScriptHash era) (MultiSig era)

Multi-signature script witness accessor function for Transactions

extractKeyHashWitnessSet ∷ ∀ (r ∷ KeyRole) era. [Credential r era] → Set (KeyHash Witness era)

getKeyCombinations ∷ Era era ⇒ MultiSig era → [[KeyHash Witness era]]

Get all valid combinations of keys for given multi signature. This is mainly useful for testing.

getKeyCombination ∷ Era era ⇒ MultiSig era → [KeyHash Witness era]

Get one possible combination of keys for multi signature script