Shelley.Spec.Ledger.Tx

Documentation

data Tx crypto

A fully formed transaction.

Constructors

Tx

Fields

_body ∷ !(TxBody crypto)

_witnessVKeySet ∷ !(Set (WitVKey crypto))

_witnessMSigMap ∷ Map (ScriptHash crypto) (MultiSig crypto)

_metadata ∷ Maybe MetaData

data TxBody crypto

A raw transaction

Constructors

TxBody

Fields

_inputs ∷ !(Set (TxIn crypto))

_outputs ∷ Seq (TxOut crypto)

_certs ∷ Seq (DCert crypto)

_wdrls ∷ Wdrl crypto

_txfee ∷ Coin

_ttl ∷ SlotNo

_txUpdate ∷ Maybe (Update crypto)

_mdHash ∷ Maybe (MetaDataHash crypto)

data TxOut crypto

The output of a UTxO.

Constructors

TxOut (Addr crypto) Coin

data TxIn crypto

The input of a UTxO.

Constructors

TxIn (TxId crypto) Natural

newtype TxId crypto

A unique ID of a transaction, which is computable from the transaction.

Constructors

TxId

Fields

_TxId ∷ Hash (HASH crypto) (TxBody crypto)

data WitVKey crypto

Proof/Witness that a transaction is authorized by the given key holder.

Constructors

WitVKey (VKey crypto) !(Sig crypto (TxBody crypto))

class (Crypto crypto, ToCBOR a) ⇒ MultiSignatureScript a crypto

Typeclass for multis-signature script data types. Allows for script validation and hashing.

Minimal complete definition

validateScript, hashScript

validateScript ∷ MultiSignatureScript a crypto ⇒ a → Tx crypto → Bool

hashScript ∷ MultiSignatureScript a crypto ⇒ a → ScriptHash crypto

txwitsScript ∷ Tx crypto → Map (ScriptHash crypto) (MultiSig crypto)

Multi-signature script witness accessor function for Transactions

extractKeyHash ∷ [Credential crypto] → [AnyKeyHash crypto]

extractScriptHash ∷ [Credential crypto] → [ScriptHash crypto]

extractGenKeyHash ∷ [GenKeyHash crypto] → [AnyKeyHash crypto]

getKeyCombinations ∷ MultiSig crypto → [[AnyKeyHash crypto]]

Get all valid combinations of keys for given multi signature. This is mainly useful for testing.

getKeyCombination ∷ MultiSig crypto → [AnyKeyHash crypto]

Get one possible combination of keys for multi signature script

txToCBORWits ∷ Tx crypto → CBORWits crypto

cborWitsToTx ∷ Crypto crypto ⇒ TxBody crypto → CBORWits crypto → Maybe MetaData → Tx crypto