Shelley.Spec.Ledger.Tx

Documentation

data Tx crypto where

A fully formed transaction.

Bundled Patterns

pattern Tx ∷ Crypto crypto ⇒ TxBody crypto → Set (WitVKey crypto) → Map (ScriptHash crypto) (MultiSig crypto) → StrictMaybe MetaData → Tx crypto

data TxBody crypto where

A raw transaction

Bundled Patterns

pattern TxBody ∷ Crypto crypto ⇒ Set (TxIn crypto) → StrictSeq (TxOut crypto) → StrictSeq (DCert crypto) → Wdrl crypto → Coin → SlotNo → StrictMaybe (Update crypto) → StrictMaybe (MetaDataHash crypto) → TxBody crypto

data TxOut crypto

The output of a UTxO.

Constructors

TxOut !(Addr crypto) !Coin

data TxIn crypto

The input of a UTxO.

Constructors

TxIn !(TxId crypto) !Natural

newtype TxId crypto

A unique ID of a transaction, which is computable from the transaction.

Constructors

TxId

Fields

_TxId ∷ Hash crypto (TxBody crypto)

data Wits crypto

Constructors

Wits (Set (WitVKey crypto)) (MultiSigMap crypto)

decodeWits ∷ Crypto crypto ⇒ Decoder s (Annotator (Wits crypto))

segwitTx ∷ Crypto crypto ⇒ Annotator (TxBody crypto) → (Annotator (Wits crypto), Annotator LByteString) → Maybe (Annotator MetaData) → Annotator (Tx crypto)

data WitVKey crypto where

Proof/Witness that a transaction is authorized by the given key holder.

Bundled Patterns

pattern WitVKey ∷ Crypto crypto ⇒ VKey Witness crypto → SignedDSIGN crypto (Hash crypto (TxBody crypto)) → WitVKey crypto

class (Crypto crypto, ToCBOR a) ⇒ MultiSignatureScript a crypto

Typeclass for multis-signature script data types. Allows for script validation and hashing.

Minimal complete definition

validateScript, hashScript

validateScript ∷ MultiSignatureScript a crypto ⇒ a → Tx crypto → Bool

hashScript ∷ MultiSignatureScript a crypto ⇒ a → ScriptHash crypto

txwitsScript ∷ Crypto crypto ⇒ Tx crypto → Map (ScriptHash crypto) (MultiSig crypto)

Multi-signature script witness accessor function for Transactions

extractKeyHash ∷ [Credential kr crypto] → [KeyHash kr crypto]

extractScriptHash ∷ [Credential Payment crypto] → [ScriptHash crypto]

getKeyCombinations ∷ Crypto crypto ⇒ MultiSig crypto → [[KeyHash Witness crypto]]

Get all valid combinations of keys for given multi signature. This is mainly useful for testing.

getKeyCombination ∷ Crypto crypto ⇒ MultiSig crypto → [KeyHash Witness crypto]

Get one possible combination of keys for multi signature script