Shelley.Spec.Ledger.TxData

Documentation

data Delegation crypto

The delegation of one stake key to another.

Constructors

Delegation

Fields

_delegator ∷ Credential crypto

_delegatee ∷ KeyHash crypto

newtype Url

Constructors

Url

Fields

unUrl ∷ Text64

data PoolMetaData

Constructors

PoolMetaData

Fields

_poolMDUrl ∷ Url

_poolMDHash ∷ ByteString

data PoolParams crypto

A stake pool.

Constructors

PoolParams

Fields

_poolPubKey ∷ KeyHash crypto

_poolVrf ∷ Hash (HASH crypto) (VerKeyVRF (VRF crypto))

_poolPledge ∷ Coin

_poolCost ∷ Coin

_poolMargin ∷ UnitInterval

_poolRAcnt ∷ RewardAcnt crypto

_poolOwners ∷ Set (KeyHash crypto)

_poolRelays ∷ Seq Url

_poolMD ∷ Maybe PoolMetaData

newtype RewardAcnt crypto

An account based address for rewards

Constructors

RewardAcnt

Fields

getRwdCred ∷ Credential crypto

data Credential crypto

Script hash or key hash for a payment or a staking object.

Constructors

ScriptHashObj (ScriptHash crypto)

KeyHashObj (KeyHash crypto)

newtype GenesisCredential crypto

Constructors

GenesisCredential (GenKeyHash crypto)

data Addr crypto

An address for UTxO.

Constructors

AddrBase (Credential crypto) (Credential crypto)

AddrEnterprise (Credential crypto)

AddrPtr (Credential crypto) Ptr

AddrBootstrap (KeyHash crypto)

type Ix = Natural

data Ptr

Pointer to a slot, transaction index and index in certificate list.

Constructors

Ptr SlotNo Ix Ix

newtype Wdrl crypto

Constructors

Wdrl

Fields

unWdrl ∷ Map (RewardAcnt crypto) Coin

newtype TxId crypto

A unique ID of a transaction, which is computable from the transaction.

Constructors

TxId

Fields

_TxId ∷ Hash (HASH crypto) (TxBody crypto)

data TxIn crypto

The input of a UTxO.

Constructors

TxIn (TxId crypto) Natural

data TxOut crypto

The output of a UTxO.

Constructors

TxOut (Addr crypto) Coin

data DelegCert crypto

Constructors

RegKey (Credential crypto)
A stake key registration certificate.

DeRegKey (Credential crypto)
A stake key deregistration certificate.

Delegate (Delegation crypto)
A stake delegation certificate.

data PoolCert crypto

Constructors

RegPool (PoolParams crypto)
A stake pool registration certificate.

RetirePool (KeyHash crypto) EpochNo
A stake pool retirement certificate.

newtype GenesisDelegate crypto

Genesis key delegation certificate

Constructors

GenesisDelegate (GenKeyHash crypto, KeyHash crypto)

newtype MIRCert crypto

Move instantaneous rewards certificate

Constructors

MIRCert (Map (Credential crypto) Coin)

data DCert crypto

A heavyweight certificate.

Constructors

DCertDeleg (DelegCert crypto)

DCertPool (PoolCert crypto)

DCertGenesis (GenesisDelegate crypto)

DCertMir (MIRCert crypto)

data TxBody crypto

A raw transaction

Constructors

TxBody

Fields

_inputs ∷ !(Set (TxIn crypto))

_outputs ∷ Seq (TxOut crypto)

_certs ∷ Seq (DCert crypto)

_wdrls ∷ Wdrl crypto

_txfee ∷ Coin

_ttl ∷ SlotNo

_txUpdate ∷ Maybe (Update crypto)

_mdHash ∷ Maybe (MetaDataHash crypto)

data WitVKey crypto

Proof/Witness that a transaction is authorized by the given key holder.

Constructors

WitVKey (VKey crypto) !(Sig crypto (TxBody crypto))

witKeyHash ∷ ∀ crypto. Crypto crypto ⇒ WitVKey crypto → AnyKeyHash crypto

newtype StakeCreds crypto

Constructors

StakeCreds (Map (Credential crypto) SlotNo)

addStakeCreds ∷ Credential crypto → SlotNo → StakeCreds crypto → StakeCreds crypto

newtype StakePools crypto

Constructors

StakePools

Fields

unStakePools ∷ Map (KeyHash crypto) SlotNo