Shelley.Spec.Ledger.UTxO

Description

This module defines the types and functions for a simple UTxO Ledger as specified in A Simplified Formal Specification of a UTxO Ledger.

Primitives

newtype UTxO crypto

The unspent transaction outputs.

Constructors

UTxO (Map (TxIn crypto) (TxOut crypto))

Functions

hashTxBody ∷ Crypto crypto ⇒ TxBody crypto → Hash crypto (TxBody crypto)

Compute the hash of a transaction body.

txid ∷ Crypto crypto ⇒ TxBody crypto → TxId crypto

Compute the id of a transaction.

txins ∷ Crypto crypto ⇒ TxBody crypto → Set (TxIn crypto)

Compute the UTxO inputs of a transaction.

txinLookup ∷ TxIn crypto → UTxO crypto → Maybe (TxOut crypto)

Lookup a txin for a given UTxO collection

txouts ∷ Crypto crypto ⇒ TxBody crypto → UTxO crypto

Compute the transaction outputs of a transaction.

txup ∷ Crypto crypto ⇒ Tx crypto → Maybe (Update crypto)

balance ∷ UTxO crypto → Coin

Determine the total balance contained in the UTxO.

totalDeposits ∷ PParams → StakePools crypto → [DCert crypto] → Coin

Determine the total deposit amount needed

makeWitnessVKey ∷ ∀ crypto kr. (Crypto crypto, DSignable crypto (Hash crypto (TxBody crypto))) ⇒ Hash crypto (TxBody crypto) → KeyPair kr crypto → WitVKey crypto

Create a witness for transaction

makeWitnessesVKey ∷ (Crypto crypto, DSignable crypto (Hash crypto (TxBody crypto))) ⇒ Hash crypto (TxBody crypto) → [KeyPair a crypto] → Set (WitVKey crypto)

Create witnesses for transaction

makeWitnessesFromScriptKeys ∷ (Crypto crypto, DSignable crypto (Hash crypto (TxBody crypto))) ⇒ Hash crypto (TxBody crypto) → Map (KeyHash kr crypto) (KeyPair kr crypto) → Set (KeyHash kr crypto) → Set (WitVKey crypto)

From a list of key pairs and a set of key hashes required for a multi-sig scripts, return the set of required keys.

verifyWitVKey ∷ (Crypto crypto, DSignable crypto (Hash crypto (TxBody crypto))) ⇒ Hash crypto (TxBody crypto) → WitVKey crypto → Bool

Verify a transaction body witness

scriptsNeeded ∷ Crypto crypto ⇒ UTxO crypto → Tx crypto → Set (ScriptHash crypto)

Computes the set of script hashes required to unlock the transcation inputs and the withdrawals.

txinsScript ∷ Set (TxIn crypto) → UTxO crypto → Set (TxIn crypto)

Compute the subset of inputs of the set txInps for which each input is locked by a script in the UTxO u.